Formatted closer to Django REST API best practices.
