+++ book_recommender_react_project/src/utilities.jsx
@@ -2,7 +2,7 @@ import axios from 'axios'
 
 /** User Authentication */
 export const signUp = async (email, password) => {
-  let response = await axios.post('/api/sign_up/', {
+  let response = await axios.post('/api/user_authentication/', {
     'email' : email,
     'password' : password
   })
@@ -11,7 +11,7 @@ export const signUp = async (email, password) => {
 }
 
 export const signIn = async (email, password, setUser) => {
-  let response = await axios.post('/api/sign_in/', {
+  let response = await axios.put('/api/user_authentication/', {
     'email' : email,
     'password' : password
   })
@@ -21,12 +21,12 @@ export const signIn = async (email, password, setUser) => {
 }
 
 export const currentUser = async () => {
-  let response = await axios.get('/api/current_user/')
+  let response = await axios.get('api/user_authentication/')
   return response.data
 }
 
 export const signOut = async () => {
-  let response = await axios.post('/api/sign_out/')
+  let response = await axios.post('api/user_authentication/')
   if (response.data.logout) {
     setUser(null)
 }
@@ -35,7 +35,7 @@ export const signOut = async () => {
 /** OpenAI API */
 export const recommendations = async (quote, user_email) => {
   try {
-    let response = await axios.post('/api/recommendations/', {
+    let response = await axios.post('/api/user_recommendation/', {
       "quote" : quote,
       "user_email" : user_email
     })
@@ -50,7 +50,7 @@ export const recommendations = async (quote, user_email) => {
 /** Grabbing DB data for FE */
 export const getRecHistory = async (user_email) => {
   try {
-    const response = await axios.get('/api/user_recommendation_history/', {
+    const response = await axios.get('/api/user_recommendation/', {
       params: {
         "user_email": user_email
       }
@@ -66,7 +66,7 @@ export const getRecHistory = async (user_email) => {
 /** Deleting a recommendation */
 export const deleteRecommendation = async (recommendation_pk) => {
   console.log(recommendation_pk)
-  const response = await axios.post('api/delete_recommendation/', {
+  const response = await axios.post('api/user_recommendation/', {
     "recommendation_pk" : recommendation_pk
   })
   // console.log(response.data)
@@ -75,7 +75,7 @@ export const deleteRecommendation = async (recommendation_pk) => {
 
 /** Deleting a quote */
 export const deleteQuote = async (quote_pk) => {
-  const response = await axios.post('api/delete_quote/', {
+  const response = await axios.post('api/user_recommendation/', {
     "quote_pk": quote_pk,
   });
   return response.data
