+++ static/assets/index-36d96ced.js
@@ -66,11 +66,11 @@ Error generating stack: `+i.message+`
  * @license MIT
  */function ss(){return ss=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},ss.apply(this,arguments)}function f0(e,t){return Py({basename:t==null?void 0:t.basename,future:t==null?void 0:t.future,history:Yv({window:t==null?void 0:t.window}),hydrationData:(t==null?void 0:t.hydrationData)||d0(),routes:e,detectErrorBoundary:c0}).initialize()}function d0(){var e;let t=(e=window)==null?void 0:e.__staticRouterHydrationData;return t&&t.errors&&(t=ss({},t,{errors:p0(t.errors)})),t}function p0(e){if(!e)return null;let t=Object.entries(e),n={};for(let[r,o]of t)if(o&&o.__type==="RouteErrorResponse")n[r]=new mu(o.status,o.statusText,o.data,o.internal===!0);else if(o&&o.__type==="Error"){let i=new Error(o.message);i.stack="",n[r]=i}else n[r]=o;return n}var tf;(function(e){e.UseScrollRestoration="useScrollRestoration",e.UseSubmitImpl="useSubmitImpl",e.UseFetcher="useFetcher"})(tf||(tf={}));var nf;(function(e){e.UseFetchers="useFetchers",e.UseScrollRestoration="useScrollRestoration"})(nf||(nf={}));const gl=N.createContext(null);function $p(e,t){return function(){return e.apply(t,arguments)}}const{toString:Hp}=Object.prototype,{getPrototypeOf:Eu}=Object,Su=(e=>t=>{const n=Hp.call(t);return e[n]||(e[n]=n.slice(8,-1).toLowerCase())})(Object.create(null)),bt=e=>(e=e.toLowerCase(),t=>Su(t)===e),wl=e=>t=>typeof t===e,{isArray:Tr}=Array,No=wl("undefined");function h0(e){return e!==null&&!No(e)&&e.constructor!==null&&!No(e.constructor)&&yn(e.constructor.isBuffer)&&e.constructor.isBuffer(e)}const Vp=bt("ArrayBuffer");function m0(e){let t;return typeof ArrayBuffer<"u"&&ArrayBuffer.isView?t=ArrayBuffer.isView(e):t=e&&e.buffer&&Vp(e.buffer),t}const v0=wl("string"),yn=wl("function"),Wp=wl("number"),xu=e=>e!==null&&typeof e=="object",y0=e=>e===!0||e===!1,Ei=e=>{if(Su(e)!=="object")return!1;const t=Eu(e);return(t===null||t===Object.prototype||Object.getPrototypeOf(t)===null)&&!(Symbol.toStringTag in e)&&!(Symbol.iterator in e)},g0=bt("Date"),w0=bt("File"),E0=bt("Blob"),S0=bt("FileList"),x0=e=>xu(e)&&yn(e.pipe),_0=e=>{const t="[object FormData]";return e&&(typeof FormData=="function"&&e instanceof FormData||Hp.call(e)===t||yn(e.toString)&&e.toString()===t)},k0=bt("URLSearchParams"),C0=e=>e.trim?e.trim():e.replace(/^[\s\uFEFF\xA0]+|[\s\uFEFF\xA0]+$/g,"");function jo(e,t,{allOwnKeys:n=!1}={}){if(e===null||typeof e>"u")return;let r,o;if(typeof e!="object"&&(e=[e]),Tr(e))for(r=0,o=e.length;r<o;r++)t.call(null,e[r],r,e);else{const i=n?Object.getOwnPropertyNames(e):Object.keys(e),l=i.length;let a;for(r=0;r<l;r++)a=i[r],t.call(null,e[a],a,e)}}function bp(e,t){t=t.toLowerCase();const n=Object.keys(e);let r=n.length,o;for(;r-- >0;)if(o=n[r],t===o.toLowerCase())return o;return null}const Qp=(()=>typeof globalThis<"u"?globalThis:typeof self<"u"?self:typeof window<"u"?window:global)(),Kp=e=>!No(e)&&e!==Qp;function us(){const{caseless:e}=Kp(this)&&this||{},t={},n=(r,o)=>{const i=e&&bp(t,o)||o;Ei(t[i])&&Ei(r)?t[i]=us(t[i],r):Ei(r)?t[i]=us({},r):Tr(r)?t[i]=r.slice():t[i]=r};for(let r=0,o=arguments.length;r<o;r++)arguments[r]&&jo(arguments[r],n);return t}const R0=(e,t,n,{allOwnKeys:r}={})=>(jo(t,(o,i)=>{n&&yn(o)?e[i]=$p(o,n):e[i]=o},{allOwnKeys:r}),e),P0=e=>(e.charCodeAt(0)===65279&&(e=e.slice(1)),e),O0=(e,t,n,r)=>{e.prototype=Object.create(t.prototype,r),e.prototype.constructor=e,Object.defineProperty(e,"super",{value:t.prototype}),n&&Object.assign(e.prototype,n)},N0=(e,t,n,r)=>{let o,i,l;const a={};if(t=t||{},e==null)return t;do{for(o=Object.getOwnPropertyNames(e),i=o.length;i-- >0;)l=o[i],(!r||r(l,e,t))&&!a[l]&&(t[l]=e[l],a[l]=!0);e=n!==!1&&Eu(e)}while(e&&(!n||n(e,t))&&e!==Object.prototype);return t},T0=(e,t,n)=>{e=String(e),(n===void 0||n>e.length)&&(n=e.length),n-=t.length;const r=e.indexOf(t,n);return r!==-1&&r===n},L0=e=>{if(!e)return null;if(Tr(e))return e;let t=e.length;if(!Wp(t))return null;const n=new Array(t);for(;t-- >0;)n[t]=e[t];return n},D0=(e=>t=>e&&t instanceof e)(typeof Uint8Array<"u"&&Eu(Uint8Array)),A0=(e,t)=>{const r=(e&&e[Symbol.iterator]).call(e);let o;for(;(o=r.next())&&!o.done;){const i=o.value;t.call(e,i[0],i[1])}},M0=(e,t)=>{let n;const r=[];for(;(n=e.exec(t))!==null;)r.push(n);return r},j0=bt("HTMLFormElement"),z0=e=>e.toLowerCase().replace(/[-_\s]([a-z\d])(\w*)/g,function(n,r,o){return r.toUpperCase()+o}),rf=(({hasOwnProperty:e})=>(t,n)=>e.call(t,n))(Object.prototype),F0=bt("RegExp"),Gp=(e,t)=>{const n=Object.getOwnPropertyDescriptors(e),r={};jo(n,(o,i)=>{t(o,i,e)!==!1&&(r[i]=o)}),Object.defineProperties(e,r)},U0=e=>{Gp(e,(t,n)=>{if(yn(e)&&["arguments","caller","callee"].indexOf(n)!==-1)return!1;const r=e[n];if(yn(r)){if(t.enumerable=!1,"writable"in t){t.writable=!1;return}t.set||(t.set=()=>{throw Error("Can not rewrite read-only method '"+n+"'")})}})},I0=(e,t)=>{const n={},r=o=>{o.forEach(i=>{n[i]=!0})};return Tr(e)?r(e):r(String(e).split(t)),n},B0=()=>{},$0=(e,t)=>(e=+e,Number.isFinite(e)?e:t),ia="abcdefghijklmnopqrstuvwxyz",of="0123456789",Yp={DIGIT:of,ALPHA:ia,ALPHA_DIGIT:ia+ia.toUpperCase()+of},H0=(e=16,t=Yp.ALPHA_DIGIT)=>{let n="";const{length:r}=t;for(;e--;)n+=t[Math.random()*r|0];return n};function V0(e){return!!(e&&yn(e.append)&&e[Symbol.toStringTag]==="FormData"&&e[Symbol.iterator])}const W0=e=>{const t=new Array(10),n=(r,o)=>{if(xu(r)){if(t.indexOf(r)>=0)return;if(!("toJSON"in r)){t[o]=r;const i=Tr(r)?[]:{};return jo(r,(l,a)=>{const u=n(l,o+1);!No(u)&&(i[a]=u)}),t[o]=void 0,i}}return r};return n(e,0)},O={isArray:Tr,isArrayBuffer:Vp,isBuffer:h0,isFormData:_0,isArrayBufferView:m0,isString:v0,isNumber:Wp,isBoolean:y0,isObject:xu,isPlainObject:Ei,isUndefined:No,isDate:g0,isFile:w0,isBlob:E0,isRegExp:F0,isFunction:yn,isStream:x0,isURLSearchParams:k0,isTypedArray:D0,isFileList:S0,forEach:jo,merge:us,extend:R0,trim:C0,stripBOM:P0,inherits:O0,toFlatObject:N0,kindOf:Su,kindOfTest:bt,endsWith:T0,toArray:L0,forEachEntry:A0,matchAll:M0,isHTMLForm:j0,hasOwnProperty:rf,hasOwnProp:rf,reduceDescriptors:Gp,freezeMethods:U0,toObjectSet:I0,toCamelCase:z0,noop:B0,toFiniteNumber:$0,findKey:bp,global:Qp,isContextDefined:Kp,ALPHABET:Yp,generateString:H0,isSpecCompliantForm:V0,toJSONObject:W0};function re(e,t,n,r,o){Error.call(this),Error.captureStackTrace?Error.captureStackTrace(this,this.constructor):this.stack=new Error().stack,this.message=e,this.name="AxiosError",t&&(this.code=t),n&&(this.config=n),r&&(this.request=r),o&&(this.response=o)}O.inherits(re,Error,{toJSON:function(){return{message:this.message,name:this.name,description:this.description,number:this.number,fileName:this.fileName,lineNumber:this.lineNumber,columnNumber:this.columnNumber,stack:this.stack,config:O.toJSONObject(this.config),code:this.code,status:this.response&&this.response.status?this.response.status:null}}});const Xp=re.prototype,Jp={};["ERR_BAD_OPTION_VALUE","ERR_BAD_OPTION","ECONNABORTED","ETIMEDOUT","ERR_NETWORK","ERR_FR_TOO_MANY_REDIRECTS","ERR_DEPRECATED","ERR_BAD_RESPONSE","ERR_BAD_REQUEST","ERR_CANCELED","ERR_NOT_SUPPORT","ERR_INVALID_URL"].forEach(e=>{Jp[e]={value:e}});Object.defineProperties(re,Jp);Object.defineProperty(Xp,"isAxiosError",{value:!0});re.from=(e,t,n,r,o,i)=>{const l=Object.create(Xp);return O.toFlatObject(e,l,function(u){return u!==Error.prototype},a=>a!=="isAxiosError"),re.call(l,e.message,t,n,r,o),l.cause=e,l.name=e.name,i&&Object.assign(l,i),l};const b0=null;function cs(e){return O.isPlainObject(e)||O.isArray(e)}function qp(e){return O.endsWith(e,"[]")?e.slice(0,-2):e}function lf(e,t,n){return e?e.concat(t).map(function(o,i){return o=qp(o),!n&&i?"["+o+"]":o}).join(n?".":""):t}function Q0(e){return O.isArray(e)&&!e.some(cs)}const K0=O.toFlatObject(O,{},null,function(t){return/^is[A-Z]/.test(t)});function El(e,t,n){if(!O.isObject(e))throw new TypeError("target must be an object");t=t||new FormData,n=O.toFlatObject(n,{metaTokens:!0,dots:!1,indexes:!1},!1,function(w,y){return!O.isUndefined(y[w])});const r=n.metaTokens,o=n.visitor||c,i=n.dots,l=n.indexes,u=(n.Blob||typeof Blob<"u"&&Blob)&&O.isSpecCompliantForm(t);if(!O.isFunction(o))throw new TypeError("visitor must be a function");function s(m){if(m===null)return"";if(O.isDate(m))return m.toISOString();if(!u&&O.isBlob(m))throw new re("Blob is not supported. Use a Buffer instead.");return O.isArrayBuffer(m)||O.isTypedArray(m)?u&&typeof Blob=="function"?new Blob([m]):Buffer.from(m):m}function c(m,w,y){let v=m;if(m&&!y&&typeof m=="object"){if(O.endsWith(w,"{}"))w=r?w:w.slice(0,-2),m=JSON.stringify(m);else if(O.isArray(m)&&Q0(m)||(O.isFileList(m)||O.endsWith(w,"[]"))&&(v=O.toArray(m)))return w=qp(w),v.forEach(function(g,E){!(O.isUndefined(g)||g===null)&&t.append(l===!0?lf([w],E,i):l===null?w:w+"[]",s(g))}),!1}return cs(m)?!0:(t.append(lf(y,w,i),s(m)),!1)}const f=[],d=Object.assign(K0,{defaultVisitor:c,convertValue:s,isVisitable:cs});function p(m,w){if(!O.isUndefined(m)){if(f.indexOf(m)!==-1)throw Error("Circular reference detected in "+w.join("."));f.push(m),O.forEach(m,function(v,h){(!(O.isUndefined(v)||v===null)&&o.call(t,v,O.isString(h)?h.trim():h,w,d))===!0&&p(v,w?w.concat(h):[h])}),f.pop()}}if(!O.isObject(e))throw new TypeError("data must be an object");return p(e),t}function af(e){const t={"!":"%21","'":"%27","(":"%28",")":"%29","~":"%7E","%20":"+","%00":"\0"};return encodeURIComponent(e).replace(/[!'()~]|%20|%00/g,function(r){return t[r]})}function _u(e,t){this._pairs=[],e&&El(e,this,t)}const Zp=_u.prototype;Zp.append=function(t,n){this._pairs.push([t,n])};Zp.toString=function(t){const n=t?function(r){return t.call(this,r,af)}:af;return this._pairs.map(function(o){return n(o[0])+"="+n(o[1])},"").join("&")};function G0(e){return encodeURIComponent(e).replace(/%3A/gi,":").replace(/%24/g,"$").replace(/%2C/gi,",").replace(/%20/g,"+").replace(/%5B/gi,"[").replace(/%5D/gi,"]")}function eh(e,t,n){if(!t)return e;const r=n&&n.encode||G0,o=n&&n.serialize;let i;if(o?i=o(t,n):i=O.isURLSearchParams(t)?t.toString():new _u(t,n).toString(r),i){const l=e.indexOf("#");l!==-1&&(e=e.slice(0,l)),e+=(e.indexOf("?")===-1?"?":"&")+i}return e}class Y0{constructor(){this.handlers=[]}use(t,n,r){return this.handlers.push({fulfilled:t,rejected:n,synchronous:r?r.synchronous:!1,runWhen:r?r.runWhen:null}),this.handlers.length-1}eject(t){this.handlers[t]&&(this.handlers[t]=null)}clear(){this.handlers&&(this.handlers=[])}forEach(t){O.forEach(this.handlers,function(r){r!==null&&t(r)})}}const sf=Y0,th={silentJSONParsing:!0,forcedJSONParsing:!0,clarifyTimeoutError:!1},X0=typeof URLSearchParams<"u"?URLSearchParams:_u,J0=typeof FormData<"u"?FormData:null,q0=typeof Blob<"u"?Blob:null,Z0=(()=>{let e;return typeof navigator<"u"&&((e=navigator.product)==="ReactNative"||e==="NativeScript"||e==="NS")?!1:typeof window<"u"&&typeof document<"u"})(),eg=(()=>typeof WorkerGlobalScope<"u"&&self instanceof WorkerGlobalScope&&typeof self.importScripts=="function")(),Ct={isBrowser:!0,classes:{URLSearchParams:X0,FormData:J0,Blob:q0},isStandardBrowserEnv:Z0,isStandardBrowserWebWorkerEnv:eg,protocols:["http","https","file","blob","url","data"]};function tg(e,t){return El(e,new Ct.classes.URLSearchParams,Object.assign({visitor:function(n,r,o,i){return Ct.isNode&&O.isBuffer(n)?(this.append(r,n.toString("base64")),!1):i.defaultVisitor.apply(this,arguments)}},t))}function ng(e){return O.matchAll(/\w+|\[(\w*)]/g,e).map(t=>t[0]==="[]"?"":t[1]||t[0])}function rg(e){const t={},n=Object.keys(e);let r;const o=n.length;let i;for(r=0;r<o;r++)i=n[r],t[i]=e[i];return t}function nh(e){function t(n,r,o,i){let l=n[i++];const a=Number.isFinite(+l),u=i>=n.length;return l=!l&&O.isArray(o)?o.length:l,u?(O.hasOwnProp(o,l)?o[l]=[o[l],r]:o[l]=r,!a):((!o[l]||!O.isObject(o[l]))&&(o[l]=[]),t(n,r,o[l],i)&&O.isArray(o[l])&&(o[l]=rg(o[l])),!a)}if(O.isFormData(e)&&O.isFunction(e.entries)){const n={};return O.forEachEntry(e,(r,o)=>{t(ng(r),o,n,0)}),n}return null}const og={"Content-Type":void 0};function ig(e,t,n){if(O.isString(e))try{return(t||JSON.parse)(e),O.trim(e)}catch(r){if(r.name!=="SyntaxError")throw r}return(n||JSON.stringify)(e)}const Sl={transitional:th,adapter:["xhr","http"],transformRequest:[function(t,n){const r=n.getContentType()||"",o=r.indexOf("application/json")>-1,i=O.isObject(t);if(i&&O.isHTMLForm(t)&&(t=new FormData(t)),O.isFormData(t))return o&&o?JSON.stringify(nh(t)):t;if(O.isArrayBuffer(t)||O.isBuffer(t)||O.isStream(t)||O.isFile(t)||O.isBlob(t))return t;if(O.isArrayBufferView(t))return t.buffer;if(O.isURLSearchParams(t))return n.setContentType("application/x-www-form-urlencoded;charset=utf-8",!1),t.toString();let a;if(i){if(r.indexOf("application/x-www-form-urlencoded")>-1)return tg(t,this.formSerializer).toString();if((a=O.isFileList(t))||r.indexOf("multipart/form-data")>-1){const u=this.env&&this.env.FormData;return El(a?{"files[]":t}:t,u&&new u,this.formSerializer)}}return i||o?(n.setContentType("application/json",!1),ig(t)):t}],transformResponse:[function(t){const n=this.transitional||Sl.transitional,r=n&&n.forcedJSONParsing,o=this.responseType==="json";if(t&&O.isString(t)&&(r&&!this.responseType||o)){const l=!(n&&n.silentJSONParsing)&&o;try{return JSON.parse(t)}catch(a){if(l)throw a.name==="SyntaxError"?re.from(a,re.ERR_BAD_RESPONSE,this,null,this.response):a}}return t}],timeout:0,xsrfCookieName:"XSRF-TOKEN",xsrfHeaderName:"X-XSRF-TOKEN",maxContentLength:-1,maxBodyLength:-1,env:{FormData:Ct.classes.FormData,Blob:Ct.classes.Blob},validateStatus:function(t){return t>=200&&t<300},headers:{common:{Accept:"application/json, text/plain, */*"}}};O.forEach(["delete","get","head"],function(t){Sl.headers[t]={}});O.forEach(["post","put","patch"],function(t){Sl.headers[t]=O.merge(og)});const ku=Sl,lg=O.toObjectSet(["age","authorization","content-length","content-type","etag","expires","from","host","if-modified-since","if-unmodified-since","last-modified","location","max-forwards","proxy-authorization","referer","retry-after","user-agent"]),ag=e=>{const t={};let n,r,o;return e&&e.split(`
 `).forEach(function(l){o=l.indexOf(":"),n=l.substring(0,o).trim().toLowerCase(),r=l.substring(o+1).trim(),!(!n||t[n]&&lg[n])&&(n==="set-cookie"?t[n]?t[n].push(r):t[n]=[r]:t[n]=t[n]?t[n]+", "+r:r)}),t},uf=Symbol("internals");function br(e){return e&&String(e).trim().toLowerCase()}function Si(e){return e===!1||e==null?e:O.isArray(e)?e.map(Si):String(e)}function sg(e){const t=Object.create(null),n=/([^\s,;=]+)\s*(?:=\s*([^,;]+))?/g;let r;for(;r=n.exec(e);)t[r[1]]=r[2];return t}function ug(e){return/^[-_a-zA-Z]+$/.test(e.trim())}function la(e,t,n,r,o){if(O.isFunction(r))return r.call(this,t,n);if(o&&(t=n),!!O.isString(t)){if(O.isString(r))return t.indexOf(r)!==-1;if(O.isRegExp(r))return r.test(t)}}function cg(e){return e.trim().toLowerCase().replace(/([a-z\d])(\w*)/g,(t,n,r)=>n.toUpperCase()+r)}function fg(e,t){const n=O.toCamelCase(" "+t);["get","set","has"].forEach(r=>{Object.defineProperty(e,r+n,{value:function(o,i,l){return this[r].call(this,t,o,i,l)},configurable:!0})})}class xl{constructor(t){t&&this.set(t)}set(t,n,r){const o=this;function i(a,u,s){const c=br(u);if(!c)throw new Error("header name must be a non-empty string");const f=O.findKey(o,c);(!f||o[f]===void 0||s===!0||s===void 0&&o[f]!==!1)&&(o[f||u]=Si(a))}const l=(a,u)=>O.forEach(a,(s,c)=>i(s,c,u));return O.isPlainObject(t)||t instanceof this.constructor?l(t,n):O.isString(t)&&(t=t.trim())&&!ug(t)?l(ag(t),n):t!=null&&i(n,t,r),this}get(t,n){if(t=br(t),t){const r=O.findKey(this,t);if(r){const o=this[r];if(!n)return o;if(n===!0)return sg(o);if(O.isFunction(n))return n.call(this,o,r);if(O.isRegExp(n))return n.exec(o);throw new TypeError("parser must be boolean|regexp|function")}}}has(t,n){if(t=br(t),t){const r=O.findKey(this,t);return!!(r&&this[r]!==void 0&&(!n||la(this,this[r],r,n)))}return!1}delete(t,n){const r=this;let o=!1;function i(l){if(l=br(l),l){const a=O.findKey(r,l);a&&(!n||la(r,r[a],a,n))&&(delete r[a],o=!0)}}return O.isArray(t)?t.forEach(i):i(t),o}clear(t){const n=Object.keys(this);let r=n.length,o=!1;for(;r--;){const i=n[r];(!t||la(this,this[i],i,t,!0))&&(delete this[i],o=!0)}return o}normalize(t){const n=this,r={};return O.forEach(this,(o,i)=>{const l=O.findKey(r,i);if(l){n[l]=Si(o),delete n[i];return}const a=t?cg(i):String(i).trim();a!==i&&delete n[i],n[a]=Si(o),r[a]=!0}),this}concat(...t){return this.constructor.concat(this,...t)}toJSON(t){const n=Object.create(null);return O.forEach(this,(r,o)=>{r!=null&&r!==!1&&(n[o]=t&&O.isArray(r)?r.join(", "):r)}),n}[Symbol.iterator](){return Object.entries(this.toJSON())[Symbol.iterator]()}toString(){return Object.entries(this.toJSON()).map(([t,n])=>t+": "+n).join(`
-`)}get[Symbol.toStringTag](){return"AxiosHeaders"}static from(t){return t instanceof this?t:new this(t)}static concat(t,...n){const r=new this(t);return n.forEach(o=>r.set(o)),r}static accessor(t){const r=(this[uf]=this[uf]={accessors:{}}).accessors,o=this.prototype;function i(l){const a=br(l);r[a]||(fg(o,l),r[a]=!0)}return O.isArray(t)?t.forEach(i):i(t),this}}xl.accessor(["Content-Type","Content-Length","Accept","Accept-Encoding","User-Agent","Authorization"]);O.freezeMethods(xl.prototype);O.freezeMethods(xl);const It=xl;function aa(e,t){const n=this||ku,r=t||n,o=It.from(r.headers);let i=r.data;return O.forEach(e,function(a){i=a.call(n,i,o.normalize(),t?t.status:void 0)}),o.normalize(),i}function rh(e){return!!(e&&e.__CANCEL__)}function zo(e,t,n){re.call(this,e??"canceled",re.ERR_CANCELED,t,n),this.name="CanceledError"}O.inherits(zo,re,{__CANCEL__:!0});function dg(e,t,n){const r=n.config.validateStatus;!n.status||!r||r(n.status)?e(n):t(new re("Request failed with status code "+n.status,[re.ERR_BAD_REQUEST,re.ERR_BAD_RESPONSE][Math.floor(n.status/100)-4],n.config,n.request,n))}const pg=Ct.isStandardBrowserEnv?function(){return{write:function(n,r,o,i,l,a){const u=[];u.push(n+"="+encodeURIComponent(r)),O.isNumber(o)&&u.push("expires="+new Date(o).toGMTString()),O.isString(i)&&u.push("path="+i),O.isString(l)&&u.push("domain="+l),a===!0&&u.push("secure"),document.cookie=u.join("; ")},read:function(n){const r=document.cookie.match(new RegExp("(^|;\\s*)("+n+")=([^;]*)"));return r?decodeURIComponent(r[3]):null},remove:function(n){this.write(n,"",Date.now()-864e5)}}}():function(){return{write:function(){},read:function(){return null},remove:function(){}}}();function hg(e){return/^([a-z][a-z\d+\-.]*:)?\/\//i.test(e)}function mg(e,t){return t?e.replace(/\/+$/,"")+"/"+t.replace(/^\/+/,""):e}function oh(e,t){return e&&!hg(t)?mg(e,t):t}const vg=Ct.isStandardBrowserEnv?function(){const t=/(msie|trident)/i.test(navigator.userAgent),n=document.createElement("a");let r;function o(i){let l=i;return t&&(n.setAttribute("href",l),l=n.href),n.setAttribute("href",l),{href:n.href,protocol:n.protocol?n.protocol.replace(/:$/,""):"",host:n.host,search:n.search?n.search.replace(/^\?/,""):"",hash:n.hash?n.hash.replace(/^#/,""):"",hostname:n.hostname,port:n.port,pathname:n.pathname.charAt(0)==="/"?n.pathname:"/"+n.pathname}}return r=o(window.location.href),function(l){const a=O.isString(l)?o(l):l;return a.protocol===r.protocol&&a.host===r.host}}():function(){return function(){return!0}}();function yg(e){const t=/^([-+\w]{1,25})(:?\/\/|:)/.exec(e);return t&&t[1]||""}function gg(e,t){e=e||10;const n=new Array(e),r=new Array(e);let o=0,i=0,l;return t=t!==void 0?t:1e3,function(u){const s=Date.now(),c=r[i];l||(l=s),n[o]=u,r[o]=s;let f=i,d=0;for(;f!==o;)d+=n[f++],f=f%e;if(o=(o+1)%e,o===i&&(i=(i+1)%e),s-l<t)return;const p=c&&s-c;return p?Math.round(d*1e3/p):void 0}}function cf(e,t){let n=0;const r=gg(50,250);return o=>{const i=o.loaded,l=o.lengthComputable?o.total:void 0,a=i-n,u=r(a),s=i<=l;n=i;const c={loaded:i,total:l,progress:l?i/l:void 0,bytes:a,rate:u||void 0,estimated:u&&l&&s?(l-i)/u:void 0,event:o};c[t?"download":"upload"]=!0,e(c)}}const wg=typeof XMLHttpRequest<"u",Eg=wg&&function(e){return new Promise(function(n,r){let o=e.data;const i=It.from(e.headers).normalize(),l=e.responseType;let a;function u(){e.cancelToken&&e.cancelToken.unsubscribe(a),e.signal&&e.signal.removeEventListener("abort",a)}O.isFormData(o)&&(Ct.isStandardBrowserEnv||Ct.isStandardBrowserWebWorkerEnv)&&i.setContentType(!1);let s=new XMLHttpRequest;if(e.auth){const p=e.auth.username||"",m=e.auth.password?unescape(encodeURIComponent(e.auth.password)):"";i.set("Authorization","Basic "+btoa(p+":"+m))}const c=oh(e.baseURL,e.url);s.open(e.method.toUpperCase(),eh(c,e.params,e.paramsSerializer),!0),s.timeout=e.timeout;function f(){if(!s)return;const p=It.from("getAllResponseHeaders"in s&&s.getAllResponseHeaders()),w={data:!l||l==="text"||l==="json"?s.responseText:s.response,status:s.status,statusText:s.statusText,headers:p,config:e,request:s};dg(function(v){n(v),u()},function(v){r(v),u()},w),s=null}if("onloadend"in s?s.onloadend=f:s.onreadystatechange=function(){!s||s.readyState!==4||s.status===0&&!(s.responseURL&&s.responseURL.indexOf("file:")===0)||setTimeout(f)},s.onabort=function(){s&&(r(new re("Request aborted",re.ECONNABORTED,e,s)),s=null)},s.onerror=function(){r(new re("Network Error",re.ERR_NETWORK,e,s)),s=null},s.ontimeout=function(){let m=e.timeout?"timeout of "+e.timeout+"ms exceeded":"timeout exceeded";const w=e.transitional||th;e.timeoutErrorMessage&&(m=e.timeoutErrorMessage),r(new re(m,w.clarifyTimeoutError?re.ETIMEDOUT:re.ECONNABORTED,e,s)),s=null},Ct.isStandardBrowserEnv){const p=(e.withCredentials||vg(c))&&e.xsrfCookieName&&pg.read(e.xsrfCookieName);p&&i.set(e.xsrfHeaderName,p)}o===void 0&&i.setContentType(null),"setRequestHeader"in s&&O.forEach(i.toJSON(),function(m,w){s.setRequestHeader(w,m)}),O.isUndefined(e.withCredentials)||(s.withCredentials=!!e.withCredentials),l&&l!=="json"&&(s.responseType=e.responseType),typeof e.onDownloadProgress=="function"&&s.addEventListener("progress",cf(e.onDownloadProgress,!0)),typeof e.onUploadProgress=="function"&&s.upload&&s.upload.addEventListener("progress",cf(e.onUploadProgress)),(e.cancelToken||e.signal)&&(a=p=>{s&&(r(!p||p.type?new zo(null,e,s):p),s.abort(),s=null)},e.cancelToken&&e.cancelToken.subscribe(a),e.signal&&(e.signal.aborted?a():e.signal.addEventListener("abort",a)));const d=yg(c);if(d&&Ct.protocols.indexOf(d)===-1){r(new re("Unsupported protocol "+d+":",re.ERR_BAD_REQUEST,e));return}s.send(o||null)})},xi={http:b0,xhr:Eg};O.forEach(xi,(e,t)=>{if(e){try{Object.defineProperty(e,"name",{value:t})}catch{}Object.defineProperty(e,"adapterName",{value:t})}});const Sg={getAdapter:e=>{e=O.isArray(e)?e:[e];const{length:t}=e;let n,r;for(let o=0;o<t&&(n=e[o],!(r=O.isString(n)?xi[n.toLowerCase()]:n));o++);if(!r)throw r===!1?new re(`Adapter ${n} is not supported by the environment`,"ERR_NOT_SUPPORT"):new Error(O.hasOwnProp(xi,n)?`Adapter '${n}' is not available in the build`:`Unknown adapter '${n}'`);if(!O.isFunction(r))throw new TypeError("adapter is not a function");return r},adapters:xi};function sa(e){if(e.cancelToken&&e.cancelToken.throwIfRequested(),e.signal&&e.signal.aborted)throw new zo(null,e)}function ff(e){return sa(e),e.headers=It.from(e.headers),e.data=aa.call(e,e.transformRequest),["post","put","patch"].indexOf(e.method)!==-1&&e.headers.setContentType("application/x-www-form-urlencoded",!1),Sg.getAdapter(e.adapter||ku.adapter)(e).then(function(r){return sa(e),r.data=aa.call(e,e.transformResponse,r),r.headers=It.from(r.headers),r},function(r){return rh(r)||(sa(e),r&&r.response&&(r.response.data=aa.call(e,e.transformResponse,r.response),r.response.headers=It.from(r.response.headers))),Promise.reject(r)})}const df=e=>e instanceof It?e.toJSON():e;function Cr(e,t){t=t||{};const n={};function r(s,c,f){return O.isPlainObject(s)&&O.isPlainObject(c)?O.merge.call({caseless:f},s,c):O.isPlainObject(c)?O.merge({},c):O.isArray(c)?c.slice():c}function o(s,c,f){if(O.isUndefined(c)){if(!O.isUndefined(s))return r(void 0,s,f)}else return r(s,c,f)}function i(s,c){if(!O.isUndefined(c))return r(void 0,c)}function l(s,c){if(O.isUndefined(c)){if(!O.isUndefined(s))return r(void 0,s)}else return r(void 0,c)}function a(s,c,f){if(f in t)return r(s,c);if(f in e)return r(void 0,s)}const u={url:i,method:i,data:i,baseURL:l,transformRequest:l,transformResponse:l,paramsSerializer:l,timeout:l,timeoutMessage:l,withCredentials:l,adapter:l,responseType:l,xsrfCookieName:l,xsrfHeaderName:l,onUploadProgress:l,onDownloadProgress:l,decompress:l,maxContentLength:l,maxBodyLength:l,beforeRedirect:l,transport:l,httpAgent:l,httpsAgent:l,cancelToken:l,socketPath:l,responseEncoding:l,validateStatus:a,headers:(s,c)=>o(df(s),df(c),!0)};return O.forEach(Object.keys(e).concat(Object.keys(t)),function(c){const f=u[c]||o,d=f(e[c],t[c],c);O.isUndefined(d)&&f!==a||(n[c]=d)}),n}const ih="1.3.4",Cu={};["object","boolean","number","function","string","symbol"].forEach((e,t)=>{Cu[e]=function(r){return typeof r===e||"a"+(t<1?"n ":" ")+e}});const pf={};Cu.transitional=function(t,n,r){function o(i,l){return"[Axios v"+ih+"] Transitional option '"+i+"'"+l+(r?". "+r:"")}return(i,l,a)=>{if(t===!1)throw new re(o(l," has been removed"+(n?" in "+n:"")),re.ERR_DEPRECATED);return n&&!pf[l]&&(pf[l]=!0,console.warn(o(l," has been deprecated since v"+n+" and will be removed in the near future"))),t?t(i,l,a):!0}};function xg(e,t,n){if(typeof e!="object")throw new re("options must be an object",re.ERR_BAD_OPTION_VALUE);const r=Object.keys(e);let o=r.length;for(;o-- >0;){const i=r[o],l=t[i];if(l){const a=e[i],u=a===void 0||l(a,i,e);if(u!==!0)throw new re("option "+i+" must be "+u,re.ERR_BAD_OPTION_VALUE);continue}if(n!==!0)throw new re("Unknown option "+i,re.ERR_BAD_OPTION)}}const fs={assertOptions:xg,validators:Cu},Xt=fs.validators;class Xi{constructor(t){this.defaults=t,this.interceptors={request:new sf,response:new sf}}request(t,n){typeof t=="string"?(n=n||{},n.url=t):n=t||{},n=Cr(this.defaults,n);const{transitional:r,paramsSerializer:o,headers:i}=n;r!==void 0&&fs.assertOptions(r,{silentJSONParsing:Xt.transitional(Xt.boolean),forcedJSONParsing:Xt.transitional(Xt.boolean),clarifyTimeoutError:Xt.transitional(Xt.boolean)},!1),o!==void 0&&fs.assertOptions(o,{encode:Xt.function,serialize:Xt.function},!0),n.method=(n.method||this.defaults.method||"get").toLowerCase();let l;l=i&&O.merge(i.common,i[n.method]),l&&O.forEach(["delete","get","head","post","put","patch","common"],m=>{delete i[m]}),n.headers=It.concat(l,i);const a=[];let u=!0;this.interceptors.request.forEach(function(w){typeof w.runWhen=="function"&&w.runWhen(n)===!1||(u=u&&w.synchronous,a.unshift(w.fulfilled,w.rejected))});const s=[];this.interceptors.response.forEach(function(w){s.push(w.fulfilled,w.rejected)});let c,f=0,d;if(!u){const m=[ff.bind(this),void 0];for(m.unshift.apply(m,a),m.push.apply(m,s),d=m.length,c=Promise.resolve(n);f<d;)c=c.then(m[f++],m[f++]);return c}d=a.length;let p=n;for(f=0;f<d;){const m=a[f++],w=a[f++];try{p=m(p)}catch(y){w.call(this,y);break}}try{c=ff.call(this,p)}catch(m){return Promise.reject(m)}for(f=0,d=s.length;f<d;)c=c.then(s[f++],s[f++]);return c}getUri(t){t=Cr(this.defaults,t);const n=oh(t.baseURL,t.url);return eh(n,t.params,t.paramsSerializer)}}O.forEach(["delete","get","head","options"],function(t){Xi.prototype[t]=function(n,r){return this.request(Cr(r||{},{method:t,url:n,data:(r||{}).data}))}});O.forEach(["post","put","patch"],function(t){function n(r){return function(i,l,a){return this.request(Cr(a||{},{method:t,headers:r?{"Content-Type":"multipart/form-data"}:{},url:i,data:l}))}}Xi.prototype[t]=n(),Xi.prototype[t+"Form"]=n(!0)});const _i=Xi;class Ru{constructor(t){if(typeof t!="function")throw new TypeError("executor must be a function.");let n;this.promise=new Promise(function(i){n=i});const r=this;this.promise.then(o=>{if(!r._listeners)return;let i=r._listeners.length;for(;i-- >0;)r._listeners[i](o);r._listeners=null}),this.promise.then=o=>{let i;const l=new Promise(a=>{r.subscribe(a),i=a}).then(o);return l.cancel=function(){r.unsubscribe(i)},l},t(function(i,l,a){r.reason||(r.reason=new zo(i,l,a),n(r.reason))})}throwIfRequested(){if(this.reason)throw this.reason}subscribe(t){if(this.reason){t(this.reason);return}this._listeners?this._listeners.push(t):this._listeners=[t]}unsubscribe(t){if(!this._listeners)return;const n=this._listeners.indexOf(t);n!==-1&&this._listeners.splice(n,1)}static source(){let t;return{token:new Ru(function(o){t=o}),cancel:t}}}const _g=Ru;function kg(e){return function(n){return e.apply(null,n)}}function Cg(e){return O.isObject(e)&&e.isAxiosError===!0}const ds={Continue:100,SwitchingProtocols:101,Processing:102,EarlyHints:103,Ok:200,Created:201,Accepted:202,NonAuthoritativeInformation:203,NoContent:204,ResetContent:205,PartialContent:206,MultiStatus:207,AlreadyReported:208,ImUsed:226,MultipleChoices:300,MovedPermanently:301,Found:302,SeeOther:303,NotModified:304,UseProxy:305,Unused:306,TemporaryRedirect:307,PermanentRedirect:308,BadRequest:400,Unauthorized:401,PaymentRequired:402,Forbidden:403,NotFound:404,MethodNotAllowed:405,NotAcceptable:406,ProxyAuthenticationRequired:407,RequestTimeout:408,Conflict:409,Gone:410,LengthRequired:411,PreconditionFailed:412,PayloadTooLarge:413,UriTooLong:414,UnsupportedMediaType:415,RangeNotSatisfiable:416,ExpectationFailed:417,ImATeapot:418,MisdirectedRequest:421,UnprocessableEntity:422,Locked:423,FailedDependency:424,TooEarly:425,UpgradeRequired:426,PreconditionRequired:428,TooManyRequests:429,RequestHeaderFieldsTooLarge:431,UnavailableForLegalReasons:451,InternalServerError:500,NotImplemented:501,BadGateway:502,ServiceUnavailable:503,GatewayTimeout:504,HttpVersionNotSupported:505,VariantAlsoNegotiates:506,InsufficientStorage:507,LoopDetected:508,NotExtended:510,NetworkAuthenticationRequired:511};Object.entries(ds).forEach(([e,t])=>{ds[t]=e});const Rg=ds;function lh(e){const t=new _i(e),n=$p(_i.prototype.request,t);return O.extend(n,_i.prototype,t,{allOwnKeys:!0}),O.extend(n,t,null,{allOwnKeys:!0}),n.create=function(o){return lh(Cr(e,o))},n}const Oe=lh(ku);Oe.Axios=_i;Oe.CanceledError=zo;Oe.CancelToken=_g;Oe.isCancel=rh;Oe.VERSION=ih;Oe.toFormData=El;Oe.AxiosError=re;Oe.Cancel=Oe.CanceledError;Oe.all=function(t){return Promise.all(t)};Oe.spread=kg;Oe.isAxiosError=Cg;Oe.mergeConfig=Cr;Oe.AxiosHeaders=It;Oe.formToJSON=e=>nh(O.isHTMLForm(e)?new FormData(e):e);Oe.HttpStatusCode=Rg;Oe.default=Oe;const Qt=Oe,Pg=async(e,t)=>{let n=await Qt.post("/api/sign_up/",{email:e,password:t});return console.log(n.data.success),n.data.success},Og=async(e,t,n)=>{let r=await Qt.post("/api/sign_in/",{email:e,password:t});return n(r.data),r.data.sign_in},Ng=async()=>(await Qt.get("/api/current_user/")).data,Tg=async()=>{(await Qt.post("/api/sign_out/")).data.logout&&setUser(null)},Lg=async(e,t)=>{try{return await Qt.post("/api/recommendations/",{quote:e,user_email:t})}catch(n){console.log(n)}},Dg=async e=>{try{return(await Qt.get("/api/user_recommendation_history/",{params:{user_email:e}})).data}catch(t){console.log(t)}},Ag=async e=>(console.log(e),(await Qt.post("api/delete_recommendation/",{recommendation_pk:e})).data),Mg=async e=>(await Qt.post("api/delete_quote/",{quote_pk:e})).data,jg=()=>{function e(n){let r=null;if(document.cookie&&document.cookie!==""){const o=document.cookie.split(";");for(let i=0;i<o.length;i++){const l=o[i].trim();if(l.substring(0,n.length+1)===n+"="){r=decodeURIComponent(l.substring(n.length+1));break}}}return r}const t=e("csrftoken");Qt.defaults.headers.common["X-CSRFToken"]=t},zg=()=>{N.useContext(gl);const e=Bp();return $("div",{children:$("button",{class:"btn btn-outline-secondary",onClick:()=>{Tg(),e("/")},children:xe("svg",{xmlns:"http://www.w3.org/2000/svg",width:"32",height:"32",fill:"currentColor",class:"bi bi-box-arrow-in-left",viewBox:"0 0 16 16",children:[$("path",{"fill-rule":"evenodd",d:"M10 3.5a.5.5 0 0 0-.5-.5h-8a.5.5 0 0 0-.5.5v9a.5.5 0 0 0 .5.5h8a.5.5 0 0 0 .5-.5v-2a.5.5 0 0 1 1 0v2A1.5 1.5 0 0 1 9.5 14h-8A1.5 1.5 0 0 1 0 12.5v-9A1.5 1.5 0 0 1 1.5 2h8A1.5 1.5 0 0 1 11 3.5v2a.5.5 0 0 1-1 0v-2z"}),$("path",{"fill-rule":"evenodd",d:"M4.146 8.354a.5.5 0 0 1 0-.708l3-3a.5.5 0 1 1 .708.708L5.707 7.5H14.5a.5.5 0 0 1 0 1H5.707l2.147 2.146a.5.5 0 0 1-.708.708l-3-3z"})]})})})},Fg=()=>{const e=wu();return $("div",{className:"banner",children:xe("div",{className:"banner-content",children:[e.pathname!=="/"&&$("div",{className:"signoutbar-container",children:$(zg,{})}),$("div",{className:"banner-text",children:"ChoiceReads"})]})})};function Ug(){const[e,t]=N.useState(null);return jg(),N.useEffect(()=>{(async()=>{t(await Ng())})()},[]),xe("div",{className:"app_root",children:[$(Fg,{}),$(gl.Provider,{value:{user:e,setUser:t},children:$(l0,{})})]})}var ps={},Ig={get exports(){return ps},set exports(e){ps=e}};/*!
+`)}get[Symbol.toStringTag](){return"AxiosHeaders"}static from(t){return t instanceof this?t:new this(t)}static concat(t,...n){const r=new this(t);return n.forEach(o=>r.set(o)),r}static accessor(t){const r=(this[uf]=this[uf]={accessors:{}}).accessors,o=this.prototype;function i(l){const a=br(l);r[a]||(fg(o,l),r[a]=!0)}return O.isArray(t)?t.forEach(i):i(t),this}}xl.accessor(["Content-Type","Content-Length","Accept","Accept-Encoding","User-Agent","Authorization"]);O.freezeMethods(xl.prototype);O.freezeMethods(xl);const It=xl;function aa(e,t){const n=this||ku,r=t||n,o=It.from(r.headers);let i=r.data;return O.forEach(e,function(a){i=a.call(n,i,o.normalize(),t?t.status:void 0)}),o.normalize(),i}function rh(e){return!!(e&&e.__CANCEL__)}function zo(e,t,n){re.call(this,e??"canceled",re.ERR_CANCELED,t,n),this.name="CanceledError"}O.inherits(zo,re,{__CANCEL__:!0});function dg(e,t,n){const r=n.config.validateStatus;!n.status||!r||r(n.status)?e(n):t(new re("Request failed with status code "+n.status,[re.ERR_BAD_REQUEST,re.ERR_BAD_RESPONSE][Math.floor(n.status/100)-4],n.config,n.request,n))}const pg=Ct.isStandardBrowserEnv?function(){return{write:function(n,r,o,i,l,a){const u=[];u.push(n+"="+encodeURIComponent(r)),O.isNumber(o)&&u.push("expires="+new Date(o).toGMTString()),O.isString(i)&&u.push("path="+i),O.isString(l)&&u.push("domain="+l),a===!0&&u.push("secure"),document.cookie=u.join("; ")},read:function(n){const r=document.cookie.match(new RegExp("(^|;\\s*)("+n+")=([^;]*)"));return r?decodeURIComponent(r[3]):null},remove:function(n){this.write(n,"",Date.now()-864e5)}}}():function(){return{write:function(){},read:function(){return null},remove:function(){}}}();function hg(e){return/^([a-z][a-z\d+\-.]*:)?\/\//i.test(e)}function mg(e,t){return t?e.replace(/\/+$/,"")+"/"+t.replace(/^\/+/,""):e}function oh(e,t){return e&&!hg(t)?mg(e,t):t}const vg=Ct.isStandardBrowserEnv?function(){const t=/(msie|trident)/i.test(navigator.userAgent),n=document.createElement("a");let r;function o(i){let l=i;return t&&(n.setAttribute("href",l),l=n.href),n.setAttribute("href",l),{href:n.href,protocol:n.protocol?n.protocol.replace(/:$/,""):"",host:n.host,search:n.search?n.search.replace(/^\?/,""):"",hash:n.hash?n.hash.replace(/^#/,""):"",hostname:n.hostname,port:n.port,pathname:n.pathname.charAt(0)==="/"?n.pathname:"/"+n.pathname}}return r=o(window.location.href),function(l){const a=O.isString(l)?o(l):l;return a.protocol===r.protocol&&a.host===r.host}}():function(){return function(){return!0}}();function yg(e){const t=/^([-+\w]{1,25})(:?\/\/|:)/.exec(e);return t&&t[1]||""}function gg(e,t){e=e||10;const n=new Array(e),r=new Array(e);let o=0,i=0,l;return t=t!==void 0?t:1e3,function(u){const s=Date.now(),c=r[i];l||(l=s),n[o]=u,r[o]=s;let f=i,d=0;for(;f!==o;)d+=n[f++],f=f%e;if(o=(o+1)%e,o===i&&(i=(i+1)%e),s-l<t)return;const p=c&&s-c;return p?Math.round(d*1e3/p):void 0}}function cf(e,t){let n=0;const r=gg(50,250);return o=>{const i=o.loaded,l=o.lengthComputable?o.total:void 0,a=i-n,u=r(a),s=i<=l;n=i;const c={loaded:i,total:l,progress:l?i/l:void 0,bytes:a,rate:u||void 0,estimated:u&&l&&s?(l-i)/u:void 0,event:o};c[t?"download":"upload"]=!0,e(c)}}const wg=typeof XMLHttpRequest<"u",Eg=wg&&function(e){return new Promise(function(n,r){let o=e.data;const i=It.from(e.headers).normalize(),l=e.responseType;let a;function u(){e.cancelToken&&e.cancelToken.unsubscribe(a),e.signal&&e.signal.removeEventListener("abort",a)}O.isFormData(o)&&(Ct.isStandardBrowserEnv||Ct.isStandardBrowserWebWorkerEnv)&&i.setContentType(!1);let s=new XMLHttpRequest;if(e.auth){const p=e.auth.username||"",m=e.auth.password?unescape(encodeURIComponent(e.auth.password)):"";i.set("Authorization","Basic "+btoa(p+":"+m))}const c=oh(e.baseURL,e.url);s.open(e.method.toUpperCase(),eh(c,e.params,e.paramsSerializer),!0),s.timeout=e.timeout;function f(){if(!s)return;const p=It.from("getAllResponseHeaders"in s&&s.getAllResponseHeaders()),w={data:!l||l==="text"||l==="json"?s.responseText:s.response,status:s.status,statusText:s.statusText,headers:p,config:e,request:s};dg(function(v){n(v),u()},function(v){r(v),u()},w),s=null}if("onloadend"in s?s.onloadend=f:s.onreadystatechange=function(){!s||s.readyState!==4||s.status===0&&!(s.responseURL&&s.responseURL.indexOf("file:")===0)||setTimeout(f)},s.onabort=function(){s&&(r(new re("Request aborted",re.ECONNABORTED,e,s)),s=null)},s.onerror=function(){r(new re("Network Error",re.ERR_NETWORK,e,s)),s=null},s.ontimeout=function(){let m=e.timeout?"timeout of "+e.timeout+"ms exceeded":"timeout exceeded";const w=e.transitional||th;e.timeoutErrorMessage&&(m=e.timeoutErrorMessage),r(new re(m,w.clarifyTimeoutError?re.ETIMEDOUT:re.ECONNABORTED,e,s)),s=null},Ct.isStandardBrowserEnv){const p=(e.withCredentials||vg(c))&&e.xsrfCookieName&&pg.read(e.xsrfCookieName);p&&i.set(e.xsrfHeaderName,p)}o===void 0&&i.setContentType(null),"setRequestHeader"in s&&O.forEach(i.toJSON(),function(m,w){s.setRequestHeader(w,m)}),O.isUndefined(e.withCredentials)||(s.withCredentials=!!e.withCredentials),l&&l!=="json"&&(s.responseType=e.responseType),typeof e.onDownloadProgress=="function"&&s.addEventListener("progress",cf(e.onDownloadProgress,!0)),typeof e.onUploadProgress=="function"&&s.upload&&s.upload.addEventListener("progress",cf(e.onUploadProgress)),(e.cancelToken||e.signal)&&(a=p=>{s&&(r(!p||p.type?new zo(null,e,s):p),s.abort(),s=null)},e.cancelToken&&e.cancelToken.subscribe(a),e.signal&&(e.signal.aborted?a():e.signal.addEventListener("abort",a)));const d=yg(c);if(d&&Ct.protocols.indexOf(d)===-1){r(new re("Unsupported protocol "+d+":",re.ERR_BAD_REQUEST,e));return}s.send(o||null)})},xi={http:b0,xhr:Eg};O.forEach(xi,(e,t)=>{if(e){try{Object.defineProperty(e,"name",{value:t})}catch{}Object.defineProperty(e,"adapterName",{value:t})}});const Sg={getAdapter:e=>{e=O.isArray(e)?e:[e];const{length:t}=e;let n,r;for(let o=0;o<t&&(n=e[o],!(r=O.isString(n)?xi[n.toLowerCase()]:n));o++);if(!r)throw r===!1?new re(`Adapter ${n} is not supported by the environment`,"ERR_NOT_SUPPORT"):new Error(O.hasOwnProp(xi,n)?`Adapter '${n}' is not available in the build`:`Unknown adapter '${n}'`);if(!O.isFunction(r))throw new TypeError("adapter is not a function");return r},adapters:xi};function sa(e){if(e.cancelToken&&e.cancelToken.throwIfRequested(),e.signal&&e.signal.aborted)throw new zo(null,e)}function ff(e){return sa(e),e.headers=It.from(e.headers),e.data=aa.call(e,e.transformRequest),["post","put","patch"].indexOf(e.method)!==-1&&e.headers.setContentType("application/x-www-form-urlencoded",!1),Sg.getAdapter(e.adapter||ku.adapter)(e).then(function(r){return sa(e),r.data=aa.call(e,e.transformResponse,r),r.headers=It.from(r.headers),r},function(r){return rh(r)||(sa(e),r&&r.response&&(r.response.data=aa.call(e,e.transformResponse,r.response),r.response.headers=It.from(r.response.headers))),Promise.reject(r)})}const df=e=>e instanceof It?e.toJSON():e;function Cr(e,t){t=t||{};const n={};function r(s,c,f){return O.isPlainObject(s)&&O.isPlainObject(c)?O.merge.call({caseless:f},s,c):O.isPlainObject(c)?O.merge({},c):O.isArray(c)?c.slice():c}function o(s,c,f){if(O.isUndefined(c)){if(!O.isUndefined(s))return r(void 0,s,f)}else return r(s,c,f)}function i(s,c){if(!O.isUndefined(c))return r(void 0,c)}function l(s,c){if(O.isUndefined(c)){if(!O.isUndefined(s))return r(void 0,s)}else return r(void 0,c)}function a(s,c,f){if(f in t)return r(s,c);if(f in e)return r(void 0,s)}const u={url:i,method:i,data:i,baseURL:l,transformRequest:l,transformResponse:l,paramsSerializer:l,timeout:l,timeoutMessage:l,withCredentials:l,adapter:l,responseType:l,xsrfCookieName:l,xsrfHeaderName:l,onUploadProgress:l,onDownloadProgress:l,decompress:l,maxContentLength:l,maxBodyLength:l,beforeRedirect:l,transport:l,httpAgent:l,httpsAgent:l,cancelToken:l,socketPath:l,responseEncoding:l,validateStatus:a,headers:(s,c)=>o(df(s),df(c),!0)};return O.forEach(Object.keys(e).concat(Object.keys(t)),function(c){const f=u[c]||o,d=f(e[c],t[c],c);O.isUndefined(d)&&f!==a||(n[c]=d)}),n}const ih="1.3.4",Cu={};["object","boolean","number","function","string","symbol"].forEach((e,t)=>{Cu[e]=function(r){return typeof r===e||"a"+(t<1?"n ":" ")+e}});const pf={};Cu.transitional=function(t,n,r){function o(i,l){return"[Axios v"+ih+"] Transitional option '"+i+"'"+l+(r?". "+r:"")}return(i,l,a)=>{if(t===!1)throw new re(o(l," has been removed"+(n?" in "+n:"")),re.ERR_DEPRECATED);return n&&!pf[l]&&(pf[l]=!0,console.warn(o(l," has been deprecated since v"+n+" and will be removed in the near future"))),t?t(i,l,a):!0}};function xg(e,t,n){if(typeof e!="object")throw new re("options must be an object",re.ERR_BAD_OPTION_VALUE);const r=Object.keys(e);let o=r.length;for(;o-- >0;){const i=r[o],l=t[i];if(l){const a=e[i],u=a===void 0||l(a,i,e);if(u!==!0)throw new re("option "+i+" must be "+u,re.ERR_BAD_OPTION_VALUE);continue}if(n!==!0)throw new re("Unknown option "+i,re.ERR_BAD_OPTION)}}const fs={assertOptions:xg,validators:Cu},Xt=fs.validators;class Xi{constructor(t){this.defaults=t,this.interceptors={request:new sf,response:new sf}}request(t,n){typeof t=="string"?(n=n||{},n.url=t):n=t||{},n=Cr(this.defaults,n);const{transitional:r,paramsSerializer:o,headers:i}=n;r!==void 0&&fs.assertOptions(r,{silentJSONParsing:Xt.transitional(Xt.boolean),forcedJSONParsing:Xt.transitional(Xt.boolean),clarifyTimeoutError:Xt.transitional(Xt.boolean)},!1),o!==void 0&&fs.assertOptions(o,{encode:Xt.function,serialize:Xt.function},!0),n.method=(n.method||this.defaults.method||"get").toLowerCase();let l;l=i&&O.merge(i.common,i[n.method]),l&&O.forEach(["delete","get","head","post","put","patch","common"],m=>{delete i[m]}),n.headers=It.concat(l,i);const a=[];let u=!0;this.interceptors.request.forEach(function(w){typeof w.runWhen=="function"&&w.runWhen(n)===!1||(u=u&&w.synchronous,a.unshift(w.fulfilled,w.rejected))});const s=[];this.interceptors.response.forEach(function(w){s.push(w.fulfilled,w.rejected)});let c,f=0,d;if(!u){const m=[ff.bind(this),void 0];for(m.unshift.apply(m,a),m.push.apply(m,s),d=m.length,c=Promise.resolve(n);f<d;)c=c.then(m[f++],m[f++]);return c}d=a.length;let p=n;for(f=0;f<d;){const m=a[f++],w=a[f++];try{p=m(p)}catch(y){w.call(this,y);break}}try{c=ff.call(this,p)}catch(m){return Promise.reject(m)}for(f=0,d=s.length;f<d;)c=c.then(s[f++],s[f++]);return c}getUri(t){t=Cr(this.defaults,t);const n=oh(t.baseURL,t.url);return eh(n,t.params,t.paramsSerializer)}}O.forEach(["delete","get","head","options"],function(t){Xi.prototype[t]=function(n,r){return this.request(Cr(r||{},{method:t,url:n,data:(r||{}).data}))}});O.forEach(["post","put","patch"],function(t){function n(r){return function(i,l,a){return this.request(Cr(a||{},{method:t,headers:r?{"Content-Type":"multipart/form-data"}:{},url:i,data:l}))}}Xi.prototype[t]=n(),Xi.prototype[t+"Form"]=n(!0)});const _i=Xi;class Ru{constructor(t){if(typeof t!="function")throw new TypeError("executor must be a function.");let n;this.promise=new Promise(function(i){n=i});const r=this;this.promise.then(o=>{if(!r._listeners)return;let i=r._listeners.length;for(;i-- >0;)r._listeners[i](o);r._listeners=null}),this.promise.then=o=>{let i;const l=new Promise(a=>{r.subscribe(a),i=a}).then(o);return l.cancel=function(){r.unsubscribe(i)},l},t(function(i,l,a){r.reason||(r.reason=new zo(i,l,a),n(r.reason))})}throwIfRequested(){if(this.reason)throw this.reason}subscribe(t){if(this.reason){t(this.reason);return}this._listeners?this._listeners.push(t):this._listeners=[t]}unsubscribe(t){if(!this._listeners)return;const n=this._listeners.indexOf(t);n!==-1&&this._listeners.splice(n,1)}static source(){let t;return{token:new Ru(function(o){t=o}),cancel:t}}}const _g=Ru;function kg(e){return function(n){return e.apply(null,n)}}function Cg(e){return O.isObject(e)&&e.isAxiosError===!0}const ds={Continue:100,SwitchingProtocols:101,Processing:102,EarlyHints:103,Ok:200,Created:201,Accepted:202,NonAuthoritativeInformation:203,NoContent:204,ResetContent:205,PartialContent:206,MultiStatus:207,AlreadyReported:208,ImUsed:226,MultipleChoices:300,MovedPermanently:301,Found:302,SeeOther:303,NotModified:304,UseProxy:305,Unused:306,TemporaryRedirect:307,PermanentRedirect:308,BadRequest:400,Unauthorized:401,PaymentRequired:402,Forbidden:403,NotFound:404,MethodNotAllowed:405,NotAcceptable:406,ProxyAuthenticationRequired:407,RequestTimeout:408,Conflict:409,Gone:410,LengthRequired:411,PreconditionFailed:412,PayloadTooLarge:413,UriTooLong:414,UnsupportedMediaType:415,RangeNotSatisfiable:416,ExpectationFailed:417,ImATeapot:418,MisdirectedRequest:421,UnprocessableEntity:422,Locked:423,FailedDependency:424,TooEarly:425,UpgradeRequired:426,PreconditionRequired:428,TooManyRequests:429,RequestHeaderFieldsTooLarge:431,UnavailableForLegalReasons:451,InternalServerError:500,NotImplemented:501,BadGateway:502,ServiceUnavailable:503,GatewayTimeout:504,HttpVersionNotSupported:505,VariantAlsoNegotiates:506,InsufficientStorage:507,LoopDetected:508,NotExtended:510,NetworkAuthenticationRequired:511};Object.entries(ds).forEach(([e,t])=>{ds[t]=e});const Rg=ds;function lh(e){const t=new _i(e),n=$p(_i.prototype.request,t);return O.extend(n,_i.prototype,t,{allOwnKeys:!0}),O.extend(n,t,null,{allOwnKeys:!0}),n.create=function(o){return lh(Cr(e,o))},n}const Oe=lh(ku);Oe.Axios=_i;Oe.CanceledError=zo;Oe.CancelToken=_g;Oe.isCancel=rh;Oe.VERSION=ih;Oe.toFormData=El;Oe.AxiosError=re;Oe.Cancel=Oe.CanceledError;Oe.all=function(t){return Promise.all(t)};Oe.spread=kg;Oe.isAxiosError=Cg;Oe.mergeConfig=Cr;Oe.AxiosHeaders=It;Oe.formToJSON=e=>nh(O.isHTMLForm(e)?new FormData(e):e);Oe.HttpStatusCode=Rg;Oe.default=Oe;const Qt=Oe,Pg=async(e,t)=>{let n=await Qt.post("/api/user_authentication/",{email:e,password:t});return console.log(n.data.success),n.data.success},Og=async(e,t,n)=>{let r=await Qt.put("/api/user_authentication/",{email:e,password:t});return n(r.data),r.data.sign_in},Ng=async()=>(await Qt.get("api/user_authentication/")).data,Tg=async()=>{(await Qt.post("api/user_authentication/")).data.logout&&setUser(null)},Lg=async(e,t)=>{try{return await Qt.post("/api/user_recommendation/",{quote:e,user_email:t})}catch(n){console.log(n)}},Dg=async e=>{try{return(await Qt.get("/api/user_recommendation/",{params:{user_email:e}})).data}catch(t){console.log(t)}},Ag=async e=>(console.log(e),(await Qt.post("api/user_recommendation/",{recommendation_pk:e})).data),Mg=async e=>(await Qt.post("api/user_recommendation/",{quote_pk:e})).data,jg=()=>{function e(n){let r=null;if(document.cookie&&document.cookie!==""){const o=document.cookie.split(";");for(let i=0;i<o.length;i++){const l=o[i].trim();if(l.substring(0,n.length+1)===n+"="){r=decodeURIComponent(l.substring(n.length+1));break}}}return r}const t=e("csrftoken");Qt.defaults.headers.common["X-CSRFToken"]=t},zg=()=>{N.useContext(gl);const e=Bp();return $("div",{children:$("button",{class:"btn btn-outline-secondary",onClick:()=>{Tg(),e("/")},children:xe("svg",{xmlns:"http://www.w3.org/2000/svg",width:"32",height:"32",fill:"currentColor",class:"bi bi-box-arrow-in-left",viewBox:"0 0 16 16",children:[$("path",{"fill-rule":"evenodd",d:"M10 3.5a.5.5 0 0 0-.5-.5h-8a.5.5 0 0 0-.5.5v9a.5.5 0 0 0 .5.5h8a.5.5 0 0 0 .5-.5v-2a.5.5 0 0 1 1 0v2A1.5 1.5 0 0 1 9.5 14h-8A1.5 1.5 0 0 1 0 12.5v-9A1.5 1.5 0 0 1 1.5 2h8A1.5 1.5 0 0 1 11 3.5v2a.5.5 0 0 1-1 0v-2z"}),$("path",{"fill-rule":"evenodd",d:"M4.146 8.354a.5.5 0 0 1 0-.708l3-3a.5.5 0 1 1 .708.708L5.707 7.5H14.5a.5.5 0 0 1 0 1H5.707l2.147 2.146a.5.5 0 0 1-.708.708l-3-3z"})]})})})},Fg=()=>{const e=wu();return $("div",{className:"banner",children:xe("div",{className:"banner-content",children:[e.pathname!=="/"&&$("div",{className:"signoutbar-container",children:$(zg,{})}),$("div",{className:"banner-text",children:"ChoiceReads"})]})})};function Ug(){const[e,t]=N.useState(null);return jg(),N.useEffect(()=>{(async()=>{t(await Ng())})()},[]),xe("div",{className:"app_root",children:[$(Fg,{}),$(gl.Provider,{value:{user:e,setUser:t},children:$(l0,{})})]})}var ps={},Ig={get exports(){return ps},set exports(e){ps=e}};/*!
 	Copyright (c) 2018 Jed Watson.
 	Licensed under the MIT License (MIT), see
 	http://jedwatson.github.io/classnames
 */(function(e){(function(){var t={}.hasOwnProperty;function n(){for(var r=[],o=0;o<arguments.length;o++){var i=arguments[o];if(i){var l=typeof i;if(l==="string"||l==="number")r.push(i);else if(Array.isArray(i)){if(i.length){var a=n.apply(null,i);a&&r.push(a)}}else if(l==="object"){if(i.toString!==Object.prototype.toString&&!i.toString.toString().includes("[native code]")){r.push(i.toString());continue}for(var u in i)t.call(i,u)&&i[u]&&r.push(u)}}}return r.join(" ")}e.exports?(n.default=n,e.exports=n):window.classNames=n})()})(Ig);const Sn=ps;function hs(){return hs=Object.assign?Object.assign.bind():function(e){for(var t=1;t<arguments.length;t++){var n=arguments[t];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(e[r]=n[r])}return e},hs.apply(this,arguments)}function ah(e,t){if(e==null)return{};var n={},r=Object.keys(e),o,i;for(i=0;i<r.length;i++)o=r[i],!(t.indexOf(o)>=0)&&(n[o]=e[o]);return n}function hf(e){return"default"+e.charAt(0).toUpperCase()+e.substr(1)}function Bg(e){var t=$g(e,"string");return typeof t=="symbol"?t:String(t)}function $g(e,t){if(typeof e!="object"||e===null)return e;var n=e[Symbol.toPrimitive];if(n!==void 0){var r=n.call(e,t||"default");if(typeof r!="object")return r;throw new TypeError("@@toPrimitive must return a primitive value.")}return(t==="string"?String:Number)(e)}function Hg(e,t,n){var r=N.useRef(e!==void 0),o=N.useState(t),i=o[0],l=o[1],a=e!==void 0,u=r.current;return r.current=a,!a&&u&&i!==t&&l(t),[a?e:i,N.useCallback(function(s){for(var c=arguments.length,f=new Array(c>1?c-1:0),d=1;d<c;d++)f[d-1]=arguments[d];n&&n.apply(void 0,[s].concat(f)),l(s)},[n])]}function Vg(e,t){return Object.keys(t).reduce(function(n,r){var o,i=n,l=i[hf(r)],a=i[r],u=ah(i,[hf(r),r].map(Bg)),s=t[r],c=Hg(a,l,e[s]),f=c[0],d=c[1];return hs({},u,(o={},o[r]=f,o[s]=d,o))},e)}function ms(e,t){return ms=Object.setPrototypeOf?Object.setPrototypeOf.bind():function(r,o){return r.__proto__=o,r},ms(e,t)}function Wg(e,t){e.prototype=Object.create(t.prototype),e.prototype.constructor=e,ms(e,t)}const bg=["xxl","xl","lg","md","sm","xs"],Qg="xs",Kg=N.createContext({prefixes:{},breakpoints:bg,minBreakpoint:Qg});function Hn(e,t){const{prefixes:n}=N.useContext(Kg);return e||n[t]||t}function Gg(e){return e&&e.ownerDocument||document}function Yg(e){var t=Gg(e);return t&&t.defaultView||window}function Xg(e,t){return Yg(e).getComputedStyle(e,t)}var Jg=/([A-Z])/g;function qg(e){return e.replace(Jg,"-$1").toLowerCase()}var Zg=/^ms-/;function li(e){return qg(e).replace(Zg,"-ms-")}var e1=/^((translate|rotate|scale)(X|Y|Z|3d)?|matrix(3d)?|perspective|skew(X|Y)?)$/i;function t1(e){return!!(e&&e1.test(e))}function Ji(e,t){var n="",r="";if(typeof t=="string")return e.style.getPropertyValue(li(t))||Xg(e).getPropertyValue(li(t));Object.keys(t).forEach(function(o){var i=t[o];!i&&i!==0?e.style.removeProperty(li(o)):t1(o)?r+=o+"("+i+") ":n+=li(o)+": "+i+";"}),r&&(n+="transform: "+r+";"),e.style.cssText+=";"+n}const mf={disabled:!1},sh=yt.createContext(null);var n1=function(t){return t.scrollTop},qr="unmounted",Zt="exited",en="entering",Rn="entered",qi="exiting",Kt=function(e){Wg(t,e);function t(r,o){var i;i=e.call(this,r,o)||this;var l=o,a=l&&!l.isMounting?r.enter:r.appear,u;return i.appearStatus=null,r.in?a?(u=Zt,i.appearStatus=en):u=Rn:r.unmountOnExit||r.mountOnEnter?u=qr:u=Zt,i.state={status:u},i.nextCallback=null,i}t.getDerivedStateFromProps=function(o,i){var l=o.in;return l&&i.status===qr?{status:Zt}:null};var n=t.prototype;return n.componentDidMount=function(){this.updateStatus(!0,this.appearStatus)},n.componentDidUpdate=function(o){var i=null;if(o!==this.props){var l=this.state.status;this.props.in?l!==en&&l!==Rn&&(i=en):(l===en||l===Rn)&&(i=qi)}this.updateStatus(!1,i)},n.componentWillUnmount=function(){this.cancelNextCallback()},n.getTimeouts=function(){var o=this.props.timeout,i,l,a;return i=l=a=o,o!=null&&typeof o!="number"&&(i=o.exit,l=o.enter,a=o.appear!==void 0?o.appear:l),{exit:i,enter:l,appear:a}},n.updateStatus=function(o,i){if(o===void 0&&(o=!1),i!==null)if(this.cancelNextCallback(),i===en){if(this.props.unmountOnExit||this.props.mountOnEnter){var l=this.props.nodeRef?this.props.nodeRef.current:Jr.findDOMNode(this);l&&n1(l)}this.performEnter(o)}else this.performExit();else this.props.unmountOnExit&&this.state.status===Zt&&this.setState({status:qr})},n.performEnter=function(o){var i=this,l=this.props.enter,a=this.context?this.context.isMounting:o,u=this.props.nodeRef?[a]:[Jr.findDOMNode(this),a],s=u[0],c=u[1],f=this.getTimeouts(),d=a?f.appear:f.enter;if(!o&&!l||mf.disabled){this.safeSetState({status:Rn},function(){i.props.onEntered(s)});return}this.props.onEnter(s,c),this.safeSetState({status:en},function(){i.props.onEntering(s,c),i.onTransitionEnd(d,function(){i.safeSetState({status:Rn},function(){i.props.onEntered(s,c)})})})},n.performExit=function(){var o=this,i=this.props.exit,l=this.getTimeouts(),a=this.props.nodeRef?void 0:Jr.findDOMNode(this);if(!i||mf.disabled){this.safeSetState({status:Zt},function(){o.props.onExited(a)});return}this.props.onExit(a),this.safeSetState({status:qi},function(){o.props.onExiting(a),o.onTransitionEnd(l.exit,function(){o.safeSetState({status:Zt},function(){o.props.onExited(a)})})})},n.cancelNextCallback=function(){this.nextCallback!==null&&(this.nextCallback.cancel(),this.nextCallback=null)},n.safeSetState=function(o,i){i=this.setNextCallback(i),this.setState(o,i)},n.setNextCallback=function(o){var i=this,l=!0;return this.nextCallback=function(a){l&&(l=!1,i.nextCallback=null,o(a))},this.nextCallback.cancel=function(){l=!1},this.nextCallback},n.onTransitionEnd=function(o,i){this.setNextCallback(i);var l=this.props.nodeRef?this.props.nodeRef.current:Jr.findDOMNode(this),a=o==null&&!this.props.addEndListener;if(!l||a){setTimeout(this.nextCallback,0);return}if(this.props.addEndListener){var u=this.props.nodeRef?[this.nextCallback]:[l,this.nextCallback],s=u[0],c=u[1];this.props.addEndListener(s,c)}o!=null&&setTimeout(this.nextCallback,o)},n.render=function(){var o=this.state.status;if(o===qr)return null;var i=this.props,l=i.children;i.in,i.mountOnEnter,i.unmountOnExit,i.appear,i.enter,i.exit,i.timeout,i.addEndListener,i.onEnter,i.onEntering,i.onEntered,i.onExit,i.onExiting,i.onExited,i.nodeRef;var a=ah(i,["children","in","mountOnEnter","unmountOnExit","appear","enter","exit","timeout","addEndListener","onEnter","onEntering","onEntered","onExit","onExiting","onExited","nodeRef"]);return yt.createElement(sh.Provider,{value:null},typeof l=="function"?l(o,a):yt.cloneElement(yt.Children.only(l),a))},t}(yt.Component);Kt.contextType=sh;Kt.propTypes={};function Xn(){}Kt.defaultProps={in:!1,mountOnEnter:!1,unmountOnExit:!1,appear:!1,enter:!0,exit:!0,onEnter:Xn,onEntering:Xn,onEntered:Xn,onExit:Xn,onExiting:Xn,onExited:Xn};Kt.UNMOUNTED=qr;Kt.EXITED=Zt;Kt.ENTERING=en;Kt.ENTERED=Rn;Kt.EXITING=qi;const r1=!!(typeof window<"u"&&window.document&&window.document.createElement);var vs=!1,ys=!1;try{var ua={get passive(){return vs=!0},get once(){return ys=vs=!0}};r1&&(window.addEventListener("test",ua,ua),window.removeEventListener("test",ua,!0))}catch{}function o1(e,t,n,r){if(r&&typeof r!="boolean"&&!ys){var o=r.once,i=r.capture,l=n;!ys&&o&&(l=n.__once||function a(u){this.removeEventListener(t,a,i),n.call(this,u)},n.__once=l),e.addEventListener(t,l,vs?r:i)}e.addEventListener(t,n,r)}function i1(e,t,n,r){var o=r&&typeof r!="boolean"?r.capture:r;e.removeEventListener(t,n,o),n.__once&&e.removeEventListener(t,n.__once,o)}function uh(e,t,n,r){return o1(e,t,n,r),function(){i1(e,t,n,r)}}function l1(e,t,n,r){if(n===void 0&&(n=!1),r===void 0&&(r=!0),e){var o=document.createEvent("HTMLEvents");o.initEvent(t,n,r),e.dispatchEvent(o)}}function a1(e){var t=Ji(e,"transitionDuration")||"",n=t.indexOf("ms")===-1?1e3:1;return parseFloat(t)*n}function s1(e,t,n){n===void 0&&(n=5);var r=!1,o=setTimeout(function(){r||l1(e,"transitionend",!0)},t+n),i=uh(e,"transitionend",function(){r=!0},{once:!0});return function(){clearTimeout(o),i()}}function u1(e,t,n,r){n==null&&(n=a1(e)||0);var o=s1(e,n,r),i=uh(e,"transitionend",t);return function(){o(),i()}}function vf(e,t){const n=Ji(e,t)||"",r=n.indexOf("ms")===-1?1e3:1;return parseFloat(n)*r}function c1(e,t){const n=vf(e,"transitionDuration"),r=vf(e,"transitionDelay"),o=u1(e,i=>{i.target===e&&(o(),t(i))},n+r)}function Qr(...e){return e.filter(t=>t!=null).reduce((t,n)=>{if(typeof n!="function")throw new Error("Invalid Argument Type, must only provide functions, undefined, or null.");return t===null?n:function(...o){t.apply(this,o),n.apply(this,o)}},null)}function f1(e){e.offsetHeight}var yf=function(t){return!t||typeof t=="function"?t:function(n){t.current=n}};function d1(e,t){var n=yf(e),r=yf(t);return function(o){n&&n(o),r&&r(o)}}function p1(e,t){return N.useMemo(function(){return d1(e,t)},[e,t])}function h1(e){return e&&"setState"in e?Jr.findDOMNode(e):e??null}const m1=yt.forwardRef(({onEnter:e,onEntering:t,onEntered:n,onExit:r,onExiting:o,onExited:i,addEndListener:l,children:a,childRef:u,...s},c)=>{const f=N.useRef(null),d=p1(f,u),p=S=>{d(h1(S))},m=S=>x=>{S&&f.current&&S(f.current,x)},w=N.useCallback(m(e),[e]),y=N.useCallback(m(t),[t]),v=N.useCallback(m(n),[n]),h=N.useCallback(m(r),[r]),g=N.useCallback(m(o),[o]),E=N.useCallback(m(i),[i]),_=N.useCallback(m(l),[l]);return $(Kt,{ref:c,...s,onEnter:w,onEntered:v,onEntering:y,onExit:h,onExited:E,onExiting:g,addEndListener:_,nodeRef:f,children:typeof a=="function"?(S,x)=>a(S,{...x,ref:p}):yt.cloneElement(a,{ref:p})})}),v1={height:["marginTop","marginBottom"],width:["marginLeft","marginRight"]};function y1(e,t){const n=`offset${e[0].toUpperCase()}${e.slice(1)}`,r=t[n],o=v1[e];return r+parseInt(Ji(t,o[0]),10)+parseInt(Ji(t,o[1]),10)}const g1={[Zt]:"collapse",[qi]:"collapsing",[en]:"collapsing",[Rn]:"collapse show"},w1=yt.forwardRef(({onEnter:e,onEntering:t,onEntered:n,onExit:r,onExiting:o,className:i,children:l,dimension:a="height",in:u=!1,timeout:s=300,mountOnEnter:c=!1,unmountOnExit:f=!1,appear:d=!1,getDimensionValue:p=y1,...m},w)=>{const y=typeof a=="function"?a():a,v=N.useMemo(()=>Qr(S=>{S.style[y]="0"},e),[y,e]),h=N.useMemo(()=>Qr(S=>{const x=`scroll${y[0].toUpperCase()}${y.slice(1)}`;S.style[y]=`${S[x]}px`},t),[y,t]),g=N.useMemo(()=>Qr(S=>{S.style[y]=null},n),[y,n]),E=N.useMemo(()=>Qr(S=>{S.style[y]=`${p(y,S)}px`,f1(S)},r),[r,p,y]),_=N.useMemo(()=>Qr(S=>{S.style[y]=null},o),[y,o]);return $(m1,{ref:w,addEndListener:c1,...m,"aria-expanded":m.role?u:null,onEnter:v,onEntering:h,onEntered:g,onExit:E,onExiting:_,childRef:l.ref,in:u,timeout:s,mountOnEnter:c,unmountOnExit:f,appear:d,children:(S,x)=>yt.cloneElement(l,{...x,className:Sn(i,l.props.className,g1[S],y==="width"&&"collapse-horizontal")})})});function ch(e,t){return Array.isArray(e)?e.includes(t):e===t}const Fo=N.createContext({});Fo.displayName="AccordionContext";const Pu=N.forwardRef(({as:e="div",bsPrefix:t,className:n,children:r,eventKey:o,...i},l)=>{const{activeEventKey:a}=N.useContext(Fo);return t=Hn(t,"accordion-collapse"),$(w1,{ref:l,in:ch(a,o),...i,className:Sn(n,t),children:$(e,{children:N.Children.only(r)})})});Pu.displayName="AccordionCollapse";const _l=N.createContext({eventKey:""});_l.displayName="AccordionItemContext";const fh=N.forwardRef(({as:e="div",bsPrefix:t,className:n,onEnter:r,onEntering:o,onEntered:i,onExit:l,onExiting:a,onExited:u,...s},c)=>{t=Hn(t,"accordion-body");const{eventKey:f}=N.useContext(_l);return $(Pu,{eventKey:f,onEnter:r,onEntering:o,onEntered:i,onExit:l,onExiting:a,onExited:u,children:$(e,{ref:c,...s,className:Sn(n,t)})})});fh.displayName="AccordionBody";function E1(e,t){const{activeEventKey:n,onSelect:r,alwaysOpen:o}=N.useContext(Fo);return i=>{let l=e===n?null:e;o&&(Array.isArray(n)?n.includes(e)?l=n.filter(a=>a!==e):l=[...n,e]:l=[e]),r==null||r(l,i),t==null||t(i)}}const Ou=N.forwardRef(({as:e="button",bsPrefix:t,className:n,onClick:r,...o},i)=>{t=Hn(t,"accordion-button");const{eventKey:l}=N.useContext(_l),a=E1(l,r),{activeEventKey:u}=N.useContext(Fo);return e==="button"&&(o.type="button"),$(e,{ref:i,onClick:a,...o,"aria-expanded":Array.isArray(u)?u.includes(l):l===u,className:Sn(n,t,!ch(u,l)&&"collapsed")})});Ou.displayName="AccordionButton";const dh=N.forwardRef(({as:e="h2",bsPrefix:t,className:n,children:r,onClick:o,...i},l)=>(t=Hn(t,"accordion-header"),$(e,{ref:l,...i,className:Sn(n,t),children:$(Ou,{onClick:o,children:r})})));dh.displayName="AccordionHeader";const ph=N.forwardRef(({as:e="div",bsPrefix:t,className:n,eventKey:r,...o},i)=>{t=Hn(t,"accordion-item");const l=N.useMemo(()=>({eventKey:r}),[r]);return $(_l.Provider,{value:l,children:$(e,{ref:i,...o,className:Sn(n,t)})})});ph.displayName="AccordionItem";const hh=N.forwardRef((e,t)=>{const{as:n="div",activeKey:r,bsPrefix:o,className:i,onSelect:l,flush:a,alwaysOpen:u,...s}=Vg(e,{activeKey:"onSelect"}),c=Hn(o,"accordion"),f=N.useMemo(()=>({activeEventKey:r,onSelect:l,alwaysOpen:u}),[r,l,u]);return $(Fo.Provider,{value:f,children:$(n,{ref:t,...s,className:Sn(i,c,a&&`${c}-flush`)})})});hh.displayName="Accordion";const ai=Object.assign(hh,{Button:Ou,Collapse:Pu,Item:ph,Header:dh,Body:fh}),S1=({user:e,update:t,triggerUpdate:n})=>{const[r,o]=N.useState([]),i=async(a,u)=>{u.recommendations.length===1&&await Mg(u.quote.pk),await Ag(a),n()};N.useEffect(()=>{(async()=>{if(e){const u=await Dg(e.email);o(u.history)}})()},[e,t]);const l=a=>{const u=new Date(a);return`${u.getMonth()+1}-${u.getDate()}-${u.getFullYear()}`};return $("div",{className:"accordion_container",children:$(ai,{defaultActiveKey:"",children:r==null?void 0:r.map((a,u)=>xe(ai.Item,{eventKey:u.toString(),children:[$(ai.Header,{children:xe("div",{className:"accordion_header_content",children:[xe("h4",{children:['"',a.quote.quote_text,'"']}),xe("p",{children:["Created on: ",l(a.quote.created_at)]})]})}),$(ai.Body,{children:a.recommendations.map((s,c)=>xe("div",{className:"rec_box mb-3",children:[xe("div",{className:"rec_box_header",children:[xe("h5",{children:[s.title," (",s.date_published,")"]}),$("button",{className:"btn btn-light",onClick:()=>i(s.pk,a),children:"Delete"})]}),xe("p",{children:[$("strong",{children:"Author:"})," ",s.author]}),xe("p",{children:[$("strong",{children:"Summary:"})," ",s.summary]}),xe("p",{children:[$("strong",{children:"Google Books Link:"})," ",$("a",{href:s.google_books_link,target:"_blank",rel:"noopener noreferrer",children:s.google_books_link})]})]},c))})]},u))})})};var gs={},x1={get exports(){return gs},set exports(e){gs=e}};(function(e,t){(function(n,r){e.exports=r(N)})(typeof self<"u"?self:Eh,n=>(()=>{var r={7403:(a,u,s)=>{s.d(u,{default:()=>Nt});var c=s(4087),f=s.n(c);const d=function(T){return new RegExp(/<[a-z][\s\S]*>/i).test(T)},p=function(T){var A=document.createElement("div");return A.innerHTML=T,A.childNodes},m=function(T,A){return Math.floor(Math.random()*(A-T+1))+T};var w="TYPE_CHARACTER",y="REMOVE_CHARACTER",v="REMOVE_ALL",h="REMOVE_LAST_VISIBLE_NODE",g="PAUSE_FOR",E="CALL_FUNCTION",_="ADD_HTML_TAG_ELEMENT",S="CHANGE_DELETE_SPEED",x="CHANGE_DELAY",R="CHANGE_CURSOR",z="PASTE_STRING",j="HTML_TAG";function H(T,A){var M=Object.keys(T);if(Object.getOwnPropertySymbols){var F=Object.getOwnPropertySymbols(T);A&&(F=F.filter(function(ie){return Object.getOwnPropertyDescriptor(T,ie).enumerable})),M.push.apply(M,F)}return M}function J(T){for(var A=1;A<arguments.length;A++){var M=arguments[A]!=null?arguments[A]:{};A%2?H(Object(M),!0).forEach(function(F){X(T,F,M[F])}):Object.getOwnPropertyDescriptors?Object.defineProperties(T,Object.getOwnPropertyDescriptors(M)):H(Object(M)).forEach(function(F){Object.defineProperty(T,F,Object.getOwnPropertyDescriptor(M,F))})}return T}function W(T){return function(A){if(Array.isArray(A))return b(A)}(T)||function(A){if(typeof Symbol<"u"&&A[Symbol.iterator]!=null||A["@@iterator"]!=null)return Array.from(A)}(T)||function(A,M){if(A){if(typeof A=="string")return b(A,M);var F=Object.prototype.toString.call(A).slice(8,-1);return F==="Object"&&A.constructor&&(F=A.constructor.name),F==="Map"||F==="Set"?Array.from(A):F==="Arguments"||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(F)?b(A,M):void 0}}(T)||function(){throw new TypeError(`Invalid attempt to spread non-iterable instance.
 In order to be iterable, non-array objects must have a [Symbol.iterator]() method.`)}()}function b(T,A){(A==null||A>T.length)&&(A=T.length);for(var M=0,F=new Array(A);M<A;M++)F[M]=T[M];return F}function q(T,A){for(var M=0;M<A.length;M++){var F=A[M];F.enumerable=F.enumerable||!1,F.configurable=!0,"value"in F&&(F.writable=!0),Object.defineProperty(T,F.key,F)}}function X(T,A,M){return A in T?Object.defineProperty(T,A,{value:M,enumerable:!0,configurable:!0,writable:!0}):T[A]=M,T}const Nt=function(){function T(F,ie){var C=this;if(function(B,ee){if(!(B instanceof ee))throw new TypeError("Cannot call a class as a function")}(this,T),X(this,"state",{cursorAnimation:null,lastFrameTime:null,pauseUntil:null,eventQueue:[],eventLoop:null,eventLoopPaused:!1,reverseCalledEvents:[],calledEvents:[],visibleNodes:[],initialOptions:null,elements:{container:null,wrapper:document.createElement("span"),cursor:document.createElement("span")}}),X(this,"options",{strings:null,cursor:"|",delay:"natural",pauseFor:1500,deleteSpeed:"natural",loop:!1,autoStart:!1,devMode:!1,skipAddStyles:!1,wrapperClassName:"Typewriter__wrapper",cursorClassName:"Typewriter__cursor",stringSplitter:null,onCreateTextNode:null,onRemoveNode:null}),X(this,"setupWrapperElement",function(){C.state.elements.container&&(C.state.elements.wrapper.className=C.options.wrapperClassName,C.state.elements.cursor.className=C.options.cursorClassName,C.state.elements.cursor.innerHTML=C.options.cursor,C.state.elements.container.innerHTML="",C.state.elements.container.appendChild(C.state.elements.wrapper),C.state.elements.container.appendChild(C.state.elements.cursor))}),X(this,"start",function(){return C.state.eventLoopPaused=!1,C.runEventLoop(),C}),X(this,"pause",function(){return C.state.eventLoopPaused=!0,C}),X(this,"stop",function(){return C.state.eventLoop&&((0,c.cancel)(C.state.eventLoop),C.state.eventLoop=null),C}),X(this,"pauseFor",function(B){return C.addEventToQueue(g,{ms:B}),C}),X(this,"typeOutAllStrings",function(){return typeof C.options.strings=="string"?(C.typeString(C.options.strings).pauseFor(C.options.pauseFor),C):(C.options.strings.forEach(function(B){C.typeString(B).pauseFor(C.options.pauseFor).deleteAll(C.options.deleteSpeed)}),C)}),X(this,"typeString",function(B){var ee=arguments.length>1&&arguments[1]!==void 0?arguments[1]:null;if(d(B))return C.typeOutHTMLString(B,ee);if(B){var he=C.options||{},we=he.stringSplitter,se=typeof we=="function"?we(B):B.split("");C.typeCharacters(se,ee)}return C}),X(this,"pasteString",function(B){var ee=arguments.length>1&&arguments[1]!==void 0?arguments[1]:null;return d(B)?C.typeOutHTMLString(B,ee,!0):(B&&C.addEventToQueue(z,{character:B,node:ee}),C)}),X(this,"typeOutHTMLString",function(B){var ee=arguments.length>1&&arguments[1]!==void 0?arguments[1]:null,he=arguments.length>2?arguments[2]:void 0,we=p(B);if(we.length>0)for(var se=0;se<we.length;se++){var be=we[se],rt=be.innerHTML;be&&be.nodeType!==3?(be.innerHTML="",C.addEventToQueue(_,{node:be,parentNode:ee}),he?C.pasteString(rt,be):C.typeString(rt,be)):be.textContent&&(he?C.pasteString(be.textContent,ee):C.typeString(be.textContent,ee))}return C}),X(this,"deleteAll",function(){var B=arguments.length>0&&arguments[0]!==void 0?arguments[0]:"natural";return C.addEventToQueue(v,{speed:B}),C}),X(this,"changeDeleteSpeed",function(B){if(!B)throw new Error("Must provide new delete speed");return C.addEventToQueue(S,{speed:B}),C}),X(this,"changeDelay",function(B){if(!B)throw new Error("Must provide new delay");return C.addEventToQueue(x,{delay:B}),C}),X(this,"changeCursor",function(B){if(!B)throw new Error("Must provide new cursor");return C.addEventToQueue(R,{cursor:B}),C}),X(this,"deleteChars",function(B){if(!B)throw new Error("Must provide amount of characters to delete");for(var ee=0;ee<B;ee++)C.addEventToQueue(y);return C}),X(this,"callFunction",function(B,ee){if(!B||typeof B!="function")throw new Error("Callbak must be a function");return C.addEventToQueue(E,{cb:B,thisArg:ee}),C}),X(this,"typeCharacters",function(B){var ee=arguments.length>1&&arguments[1]!==void 0?arguments[1]:null;if(!B||!Array.isArray(B))throw new Error("Characters must be an array");return B.forEach(function(he){C.addEventToQueue(w,{character:he,node:ee})}),C}),X(this,"removeCharacters",function(B){if(!B||!Array.isArray(B))throw new Error("Characters must be an array");return B.forEach(function(){C.addEventToQueue(y)}),C}),X(this,"addEventToQueue",function(B,ee){var he=arguments.length>2&&arguments[2]!==void 0&&arguments[2];return C.addEventToStateProperty(B,ee,he,"eventQueue")}),X(this,"addReverseCalledEvent",function(B,ee){var he=arguments.length>2&&arguments[2]!==void 0&&arguments[2],we=C.options.loop;return we?C.addEventToStateProperty(B,ee,he,"reverseCalledEvents"):C}),X(this,"addEventToStateProperty",function(B,ee){var he=arguments.length>2&&arguments[2]!==void 0&&arguments[2],we=arguments.length>3?arguments[3]:void 0,se={eventName:B,eventArgs:ee||{}};return C.state[we]=he?[se].concat(W(C.state[we])):[].concat(W(C.state[we]),[se]),C}),X(this,"runEventLoop",function(){C.state.lastFrameTime||(C.state.lastFrameTime=Date.now());var B=Date.now(),ee=B-C.state.lastFrameTime;if(!C.state.eventQueue.length){if(!C.options.loop)return;C.state.eventQueue=W(C.state.calledEvents),C.state.calledEvents=[],C.options=J({},C.state.initialOptions)}if(C.state.eventLoop=f()(C.runEventLoop),!C.state.eventLoopPaused){if(C.state.pauseUntil){if(B<C.state.pauseUntil)return;C.state.pauseUntil=null}var he,we=W(C.state.eventQueue),se=we.shift();if(!(ee<=(he=se.eventName===h||se.eventName===y?C.options.deleteSpeed==="natural"?m(40,80):C.options.deleteSpeed:C.options.delay==="natural"?m(120,160):C.options.delay))){var be=se.eventName,rt=se.eventArgs;switch(C.logInDevMode({currentEvent:se,state:C.state,delay:he}),be){case z:case w:var Vn=rt.character,Wn=rt.node,bn=document.createTextNode(Vn),Tt=bn;C.options.onCreateTextNode&&typeof C.options.onCreateTextNode=="function"&&(Tt=C.options.onCreateTextNode(Vn,bn)),Tt&&(Wn?Wn.appendChild(Tt):C.state.elements.wrapper.appendChild(Tt)),C.state.visibleNodes=[].concat(W(C.state.visibleNodes),[{type:"TEXT_NODE",character:Vn,node:Tt}]);break;case y:we.unshift({eventName:h,eventArgs:{removingCharacterNode:!0}});break;case g:var Qn=se.eventArgs.ms;C.state.pauseUntil=Date.now()+parseInt(Qn);break;case E:var Lr=se.eventArgs,kl=Lr.cb,Uo=Lr.thisArg;kl.call(Uo,{elements:C.state.elements});break;case _:var Io=se.eventArgs,Lt=Io.node,Gt=Io.parentNode;Gt?Gt.appendChild(Lt):C.state.elements.wrapper.appendChild(Lt),C.state.visibleNodes=[].concat(W(C.state.visibleNodes),[{type:j,node:Lt,parentNode:Gt||C.state.elements.wrapper}]);break;case v:var Bo=C.state.visibleNodes,xn=rt.speed,Kn=[];xn&&Kn.push({eventName:S,eventArgs:{speed:xn,temp:!0}});for(var $o=0,Ho=Bo.length;$o<Ho;$o++)Kn.push({eventName:h,eventArgs:{removingCharacterNode:!1}});xn&&Kn.push({eventName:S,eventArgs:{speed:C.options.deleteSpeed,temp:!0}}),we.unshift.apply(we,Kn);break;case h:var Cl=se.eventArgs.removingCharacterNode;if(C.state.visibleNodes.length){var k=C.state.visibleNodes.pop(),P=k.type,L=k.node,I=k.character;C.options.onRemoveNode&&typeof C.options.onRemoveNode=="function"&&C.options.onRemoveNode({node:L,character:I}),L&&L.parentNode.removeChild(L),P===j&&Cl&&we.unshift({eventName:h,eventArgs:{}})}break;case S:C.options.deleteSpeed=se.eventArgs.speed;break;case x:C.options.delay=se.eventArgs.delay;break;case R:C.options.cursor=se.eventArgs.cursor,C.state.elements.cursor.innerHTML=se.eventArgs.cursor}C.options.loop&&(se.eventName===h||se.eventArgs&&se.eventArgs.temp||(C.state.calledEvents=[].concat(W(C.state.calledEvents),[se]))),C.state.eventQueue=we,C.state.lastFrameTime=B}}}),F)if(typeof F=="string"){var ke=document.querySelector(F);if(!ke)throw new Error("Could not find container element");this.state.elements.container=ke}else this.state.elements.container=F;ie&&(this.options=J(J({},this.options),ie)),this.state.initialOptions=J({},this.options),this.init()}var A,M;return A=T,(M=[{key:"init",value:function(){var F,ie;this.setupWrapperElement(),this.addEventToQueue(R,{cursor:this.options.cursor},!0),this.addEventToQueue(v,null,!0),!window||window.___TYPEWRITER_JS_STYLES_ADDED___||this.options.skipAddStyles||(F=".Typewriter__cursor{-webkit-animation:Typewriter-cursor 1s infinite;animation:Typewriter-cursor 1s infinite;margin-left:1px}@-webkit-keyframes Typewriter-cursor{0%{opacity:0}50%{opacity:1}100%{opacity:0}}@keyframes Typewriter-cursor{0%{opacity:0}50%{opacity:1}100%{opacity:0}}",(ie=document.createElement("style")).appendChild(document.createTextNode(F)),document.head.appendChild(ie),window.___TYPEWRITER_JS_STYLES_ADDED___=!0),this.options.autoStart===!0&&this.options.strings&&this.typeOutAllStrings().start()}},{key:"logInDevMode",value:function(F){this.options.devMode&&console.log(F)}}])&&q(A.prototype,M),Object.defineProperty(A,"prototype",{writable:!1}),T}()},8552:(a,u,s)=>{var c=s(852)(s(5639),"DataView");a.exports=c},1989:(a,u,s)=>{var c=s(1789),f=s(401),d=s(7667),p=s(1327),m=s(1866);function w(y){var v=-1,h=y==null?0:y.length;for(this.clear();++v<h;){var g=y[v];this.set(g[0],g[1])}}w.prototype.clear=c,w.prototype.delete=f,w.prototype.get=d,w.prototype.has=p,w.prototype.set=m,a.exports=w},8407:(a,u,s)=>{var c=s(7040),f=s(4125),d=s(2117),p=s(7518),m=s(4705);function w(y){var v=-1,h=y==null?0:y.length;for(this.clear();++v<h;){var g=y[v];this.set(g[0],g[1])}}w.prototype.clear=c,w.prototype.delete=f,w.prototype.get=d,w.prototype.has=p,w.prototype.set=m,a.exports=w},7071:(a,u,s)=>{var c=s(852)(s(5639),"Map");a.exports=c},3369:(a,u,s)=>{var c=s(4785),f=s(1285),d=s(6e3),p=s(9916),m=s(5265);function w(y){var v=-1,h=y==null?0:y.length;for(this.clear();++v<h;){var g=y[v];this.set(g[0],g[1])}}w.prototype.clear=c,w.prototype.delete=f,w.prototype.get=d,w.prototype.has=p,w.prototype.set=m,a.exports=w},3818:(a,u,s)=>{var c=s(852)(s(5639),"Promise");a.exports=c},8525:(a,u,s)=>{var c=s(852)(s(5639),"Set");a.exports=c},8668:(a,u,s)=>{var c=s(3369),f=s(619),d=s(2385);function p(m){var w=-1,y=m==null?0:m.length;for(this.__data__=new c;++w<y;)this.add(m[w])}p.prototype.add=p.prototype.push=f,p.prototype.has=d,a.exports=p},6384:(a,u,s)=>{var c=s(8407),f=s(7465),d=s(3779),p=s(7599),m=s(4758),w=s(4309);function y(v){var h=this.__data__=new c(v);this.size=h.size}y.prototype.clear=f,y.prototype.delete=d,y.prototype.get=p,y.prototype.has=m,y.prototype.set=w,a.exports=y},2705:(a,u,s)=>{var c=s(5639).Symbol;a.exports=c},1149:(a,u,s)=>{var c=s(5639).Uint8Array;a.exports=c},577:(a,u,s)=>{var c=s(852)(s(5639),"WeakMap");a.exports=c},4963:a=>{a.exports=function(u,s){for(var c=-1,f=u==null?0:u.length,d=0,p=[];++c<f;){var m=u[c];s(m,c,u)&&(p[d++]=m)}return p}},4636:(a,u,s)=>{var c=s(2545),f=s(5694),d=s(1469),p=s(4144),m=s(5776),w=s(6719),y=Object.prototype.hasOwnProperty;a.exports=function(v,h){var g=d(v),E=!g&&f(v),_=!g&&!E&&p(v),S=!g&&!E&&!_&&w(v),x=g||E||_||S,R=x?c(v.length,String):[],z=R.length;for(var j in v)!h&&!y.call(v,j)||x&&(j=="length"||_&&(j=="offset"||j=="parent")||S&&(j=="buffer"||j=="byteLength"||j=="byteOffset")||m(j,z))||R.push(j);return R}},2488:a=>{a.exports=function(u,s){for(var c=-1,f=s.length,d=u.length;++c<f;)u[d+c]=s[c];return u}},2908:a=>{a.exports=function(u,s){for(var c=-1,f=u==null?0:u.length;++c<f;)if(s(u[c],c,u))return!0;return!1}},8470:(a,u,s)=>{var c=s(7813);a.exports=function(f,d){for(var p=f.length;p--;)if(c(f[p][0],d))return p;return-1}},8866:(a,u,s)=>{var c=s(2488),f=s(1469);a.exports=function(d,p,m){var w=p(d);return f(d)?w:c(w,m(d))}},4239:(a,u,s)=>{var c=s(2705),f=s(9607),d=s(2333),p=c?c.toStringTag:void 0;a.exports=function(m){return m==null?m===void 0?"[object Undefined]":"[object Null]":p&&p in Object(m)?f(m):d(m)}},9454:(a,u,s)=>{var c=s(4239),f=s(7005);a.exports=function(d){return f(d)&&c(d)=="[object Arguments]"}},939:(a,u,s)=>{var c=s(2492),f=s(7005);a.exports=function d(p,m,w,y,v){return p===m||(p==null||m==null||!f(p)&&!f(m)?p!=p&&m!=m:c(p,m,w,y,d,v))}},2492:(a,u,s)=>{var c=s(6384),f=s(7114),d=s(8351),p=s(6096),m=s(4160),w=s(1469),y=s(4144),v=s(6719),h="[object Arguments]",g="[object Array]",E="[object Object]",_=Object.prototype.hasOwnProperty;a.exports=function(S,x,R,z,j,H){var J=w(S),W=w(x),b=J?g:m(S),q=W?g:m(x),X=(b=b==h?E:b)==E,Nt=(q=q==h?E:q)==E,T=b==q;if(T&&y(S)){if(!y(x))return!1;J=!0,X=!1}if(T&&!X)return H||(H=new c),J||v(S)?f(S,x,R,z,j,H):d(S,x,b,R,z,j,H);if(!(1&R)){var A=X&&_.call(S,"__wrapped__"),M=Nt&&_.call(x,"__wrapped__");if(A||M){var F=A?S.value():S,ie=M?x.value():x;return H||(H=new c),j(F,ie,R,z,H)}}return!!T&&(H||(H=new c),p(S,x,R,z,j,H))}},8458:(a,u,s)=>{var c=s(3560),f=s(5346),d=s(3218),p=s(346),m=/^\[object .+?Constructor\]$/,w=Function.prototype,y=Object.prototype,v=w.toString,h=y.hasOwnProperty,g=RegExp("^"+v.call(h).replace(/[\\^$.*+?()[\]{}|]/g,"\\$&").replace(/hasOwnProperty|(function).*?(?=\\\()| for .+?(?=\\\])/g,"$1.*?")+"$");a.exports=function(E){return!(!d(E)||f(E))&&(c(E)?g:m).test(p(E))}},8749:(a,u,s)=>{var c=s(4239),f=s(1780),d=s(7005),p={};p["[object Float32Array]"]=p["[object Float64Array]"]=p["[object Int8Array]"]=p["[object Int16Array]"]=p["[object Int32Array]"]=p["[object Uint8Array]"]=p["[object Uint8ClampedArray]"]=p["[object Uint16Array]"]=p["[object Uint32Array]"]=!0,p["[object Arguments]"]=p["[object Array]"]=p["[object ArrayBuffer]"]=p["[object Boolean]"]=p["[object DataView]"]=p["[object Date]"]=p["[object Error]"]=p["[object Function]"]=p["[object Map]"]=p["[object Number]"]=p["[object Object]"]=p["[object RegExp]"]=p["[object Set]"]=p["[object String]"]=p["[object WeakMap]"]=!1,a.exports=function(m){return d(m)&&f(m.length)&&!!p[c(m)]}},280:(a,u,s)=>{var c=s(5726),f=s(6916),d=Object.prototype.hasOwnProperty;a.exports=function(p){if(!c(p))return f(p);var m=[];for(var w in Object(p))d.call(p,w)&&w!="constructor"&&m.push(w);return m}},2545:a=>{a.exports=function(u,s){for(var c=-1,f=Array(u);++c<u;)f[c]=s(c);return f}},1717:a=>{a.exports=function(u){return function(s){return u(s)}}},4757:a=>{a.exports=function(u,s){return u.has(s)}},4429:(a,u,s)=>{var c=s(5639)["__core-js_shared__"];a.exports=c},7114:(a,u,s)=>{var c=s(8668),f=s(2908),d=s(4757);a.exports=function(p,m,w,y,v,h){var g=1&w,E=p.length,_=m.length;if(E!=_&&!(g&&_>E))return!1;var S=h.get(p),x=h.get(m);if(S&&x)return S==m&&x==p;var R=-1,z=!0,j=2&w?new c:void 0;for(h.set(p,m),h.set(m,p);++R<E;){var H=p[R],J=m[R];if(y)var W=g?y(J,H,R,m,p,h):y(H,J,R,p,m,h);if(W!==void 0){if(W)continue;z=!1;break}if(j){if(!f(m,function(b,q){if(!d(j,q)&&(H===b||v(H,b,w,y,h)))return j.push(q)})){z=!1;break}}else if(H!==J&&!v(H,J,w,y,h)){z=!1;break}}return h.delete(p),h.delete(m),z}},8351:(a,u,s)=>{var c=s(2705),f=s(1149),d=s(7813),p=s(7114),m=s(8776),w=s(1814),y=c?c.prototype:void 0,v=y?y.valueOf:void 0;a.exports=function(h,g,E,_,S,x,R){switch(E){case"[object DataView]":if(h.byteLength!=g.byteLength||h.byteOffset!=g.byteOffset)return!1;h=h.buffer,g=g.buffer;case"[object ArrayBuffer]":return!(h.byteLength!=g.byteLength||!x(new f(h),new f(g)));case"[object Boolean]":case"[object Date]":case"[object Number]":return d(+h,+g);case"[object Error]":return h.name==g.name&&h.message==g.message;case"[object RegExp]":case"[object String]":return h==g+"";case"[object Map]":var z=m;case"[object Set]":var j=1&_;if(z||(z=w),h.size!=g.size&&!j)return!1;var H=R.get(h);if(H)return H==g;_|=2,R.set(h,g);var J=p(z(h),z(g),_,S,x,R);return R.delete(h),J;case"[object Symbol]":if(v)return v.call(h)==v.call(g)}return!1}},6096:(a,u,s)=>{var c=s(8234),f=Object.prototype.hasOwnProperty;a.exports=function(d,p,m,w,y,v){var h=1&m,g=c(d),E=g.length;if(E!=c(p).length&&!h)return!1;for(var _=E;_--;){var S=g[_];if(!(h?S in p:f.call(p,S)))return!1}var x=v.get(d),R=v.get(p);if(x&&R)return x==p&&R==d;var z=!0;v.set(d,p),v.set(p,d);for(var j=h;++_<E;){var H=d[S=g[_]],J=p[S];if(w)var W=h?w(J,H,S,p,d,v):w(H,J,S,d,p,v);if(!(W===void 0?H===J||y(H,J,m,w,v):W)){z=!1;break}j||(j=S=="constructor")}if(z&&!j){var b=d.constructor,q=p.constructor;b==q||!("constructor"in d)||!("constructor"in p)||typeof b=="function"&&b instanceof b&&typeof q=="function"&&q instanceof q||(z=!1)}return v.delete(d),v.delete(p),z}},1957:(a,u,s)=>{var c=typeof s.g=="object"&&s.g&&s.g.Object===Object&&s.g;a.exports=c},8234:(a,u,s)=>{var c=s(8866),f=s(9551),d=s(3674);a.exports=function(p){return c(p,d,f)}},5050:(a,u,s)=>{var c=s(7019);a.exports=function(f,d){var p=f.__data__;return c(d)?p[typeof d=="string"?"string":"hash"]:p.map}},852:(a,u,s)=>{var c=s(8458),f=s(7801);a.exports=function(d,p){var m=f(d,p);return c(m)?m:void 0}},9607:(a,u,s)=>{var c=s(2705),f=Object.prototype,d=f.hasOwnProperty,p=f.toString,m=c?c.toStringTag:void 0;a.exports=function(w){var y=d.call(w,m),v=w[m];try{w[m]=void 0;var h=!0}catch{}var g=p.call(w);return h&&(y?w[m]=v:delete w[m]),g}},9551:(a,u,s)=>{var c=s(4963),f=s(479),d=Object.prototype.propertyIsEnumerable,p=Object.getOwnPropertySymbols,m=p?function(w){return w==null?[]:(w=Object(w),c(p(w),function(y){return d.call(w,y)}))}:f;a.exports=m},4160:(a,u,s)=>{var c=s(8552),f=s(7071),d=s(3818),p=s(8525),m=s(577),w=s(4239),y=s(346),v="[object Map]",h="[object Promise]",g="[object Set]",E="[object WeakMap]",_="[object DataView]",S=y(c),x=y(f),R=y(d),z=y(p),j=y(m),H=w;(c&&H(new c(new ArrayBuffer(1)))!=_||f&&H(new f)!=v||d&&H(d.resolve())!=h||p&&H(new p)!=g||m&&H(new m)!=E)&&(H=function(J){var W=w(J),b=W=="[object Object]"?J.constructor:void 0,q=b?y(b):"";if(q)switch(q){case S:return _;case x:return v;case R:return h;case z:return g;case j:return E}return W}),a.exports=H},7801:a=>{a.exports=function(u,s){return u==null?void 0:u[s]}},1789:(a,u,s)=>{var c=s(4536);a.exports=function(){this.__data__=c?c(null):{},this.size=0}},401:a=>{a.exports=function(u){var s=this.has(u)&&delete this.__data__[u];return this.size-=s?1:0,s}},7667:(a,u,s)=>{var c=s(4536),f=Object.prototype.hasOwnProperty;a.exports=function(d){var p=this.__data__;if(c){var m=p[d];return m==="__lodash_hash_undefined__"?void 0:m}return f.call(p,d)?p[d]:void 0}},1327:(a,u,s)=>{var c=s(4536),f=Object.prototype.hasOwnProperty;a.exports=function(d){var p=this.__data__;return c?p[d]!==void 0:f.call(p,d)}},1866:(a,u,s)=>{var c=s(4536);a.exports=function(f,d){var p=this.__data__;return this.size+=this.has(f)?0:1,p[f]=c&&d===void 0?"__lodash_hash_undefined__":d,this}},5776:a=>{var u=/^(?:0|[1-9]\d*)$/;a.exports=function(s,c){var f=typeof s;return!!(c=c??9007199254740991)&&(f=="number"||f!="symbol"&&u.test(s))&&s>-1&&s%1==0&&s<c}},7019:a=>{a.exports=function(u){var s=typeof u;return s=="string"||s=="number"||s=="symbol"||s=="boolean"?u!=="__proto__":u===null}},5346:(a,u,s)=>{var c,f=s(4429),d=(c=/[^.]+$/.exec(f&&f.keys&&f.keys.IE_PROTO||""))?"Symbol(src)_1."+c:"";a.exports=function(p){return!!d&&d in p}},5726:a=>{var u=Object.prototype;a.exports=function(s){var c=s&&s.constructor;return s===(typeof c=="function"&&c.prototype||u)}},7040:a=>{a.exports=function(){this.__data__=[],this.size=0}},4125:(a,u,s)=>{var c=s(8470),f=Array.prototype.splice;a.exports=function(d){var p=this.__data__,m=c(p,d);return!(m<0||(m==p.length-1?p.pop():f.call(p,m,1),--this.size,0))}},2117:(a,u,s)=>{var c=s(8470);a.exports=function(f){var d=this.__data__,p=c(d,f);return p<0?void 0:d[p][1]}},7518:(a,u,s)=>{var c=s(8470);a.exports=function(f){return c(this.__data__,f)>-1}},4705:(a,u,s)=>{var c=s(8470);a.exports=function(f,d){var p=this.__data__,m=c(p,f);return m<0?(++this.size,p.push([f,d])):p[m][1]=d,this}},4785:(a,u,s)=>{var c=s(1989),f=s(8407),d=s(7071);a.exports=function(){this.size=0,this.__data__={hash:new c,map:new(d||f),string:new c}}},1285:(a,u,s)=>{var c=s(5050);a.exports=function(f){var d=c(this,f).delete(f);return this.size-=d?1:0,d}},6e3:(a,u,s)=>{var c=s(5050);a.exports=function(f){return c(this,f).get(f)}},9916:(a,u,s)=>{var c=s(5050);a.exports=function(f){return c(this,f).has(f)}},5265:(a,u,s)=>{var c=s(5050);a.exports=function(f,d){var p=c(this,f),m=p.size;return p.set(f,d),this.size+=p.size==m?0:1,this}},8776:a=>{a.exports=function(u){var s=-1,c=Array(u.size);return u.forEach(function(f,d){c[++s]=[d,f]}),c}},4536:(a,u,s)=>{var c=s(852)(Object,"create");a.exports=c},6916:(a,u,s)=>{var c=s(5569)(Object.keys,Object);a.exports=c},1167:(a,u,s)=>{a=s.nmd(a);var c=s(1957),f=u&&!u.nodeType&&u,d=f&&a&&!a.nodeType&&a,p=d&&d.exports===f&&c.process,m=function(){try{return d&&d.require&&d.require("util").types||p&&p.binding&&p.binding("util")}catch{}}();a.exports=m},2333:a=>{var u=Object.prototype.toString;a.exports=function(s){return u.call(s)}},5569:a=>{a.exports=function(u,s){return function(c){return u(s(c))}}},5639:(a,u,s)=>{var c=s(1957),f=typeof self=="object"&&self&&self.Object===Object&&self,d=c||f||Function("return this")();a.exports=d},619:a=>{a.exports=function(u){return this.__data__.set(u,"__lodash_hash_undefined__"),this}},2385:a=>{a.exports=function(u){return this.__data__.has(u)}},1814:a=>{a.exports=function(u){var s=-1,c=Array(u.size);return u.forEach(function(f){c[++s]=f}),c}},7465:(a,u,s)=>{var c=s(8407);a.exports=function(){this.__data__=new c,this.size=0}},3779:a=>{a.exports=function(u){var s=this.__data__,c=s.delete(u);return this.size=s.size,c}},7599:a=>{a.exports=function(u){return this.__data__.get(u)}},4758:a=>{a.exports=function(u){return this.__data__.has(u)}},4309:(a,u,s)=>{var c=s(8407),f=s(7071),d=s(3369);a.exports=function(p,m){var w=this.__data__;if(w instanceof c){var y=w.__data__;if(!f||y.length<199)return y.push([p,m]),this.size=++w.size,this;w=this.__data__=new d(y)}return w.set(p,m),this.size=w.size,this}},346:a=>{var u=Function.prototype.toString;a.exports=function(s){if(s!=null){try{return u.call(s)}catch{}try{return s+""}catch{}}return""}},7813:a=>{a.exports=function(u,s){return u===s||u!=u&&s!=s}},5694:(a,u,s)=>{var c=s(9454),f=s(7005),d=Object.prototype,p=d.hasOwnProperty,m=d.propertyIsEnumerable,w=c(function(){return arguments}())?c:function(y){return f(y)&&p.call(y,"callee")&&!m.call(y,"callee")};a.exports=w},1469:a=>{var u=Array.isArray;a.exports=u},8612:(a,u,s)=>{var c=s(3560),f=s(1780);a.exports=function(d){return d!=null&&f(d.length)&&!c(d)}},4144:(a,u,s)=>{a=s.nmd(a);var c=s(5639),f=s(5062),d=u&&!u.nodeType&&u,p=d&&a&&!a.nodeType&&a,m=p&&p.exports===d?c.Buffer:void 0,w=(m?m.isBuffer:void 0)||f;a.exports=w},8446:(a,u,s)=>{var c=s(939);a.exports=function(f,d){return c(f,d)}},3560:(a,u,s)=>{var c=s(4239),f=s(3218);a.exports=function(d){if(!f(d))return!1;var p=c(d);return p=="[object Function]"||p=="[object GeneratorFunction]"||p=="[object AsyncFunction]"||p=="[object Proxy]"}},1780:a=>{a.exports=function(u){return typeof u=="number"&&u>-1&&u%1==0&&u<=9007199254740991}},3218:a=>{a.exports=function(u){var s=typeof u;return u!=null&&(s=="object"||s=="function")}},7005:a=>{a.exports=function(u){return u!=null&&typeof u=="object"}},6719:(a,u,s)=>{var c=s(8749),f=s(1717),d=s(1167),p=d&&d.isTypedArray,m=p?f(p):c;a.exports=m},3674:(a,u,s)=>{var c=s(4636),f=s(280),d=s(8612);a.exports=function(p){return d(p)?c(p):f(p)}},479:a=>{a.exports=function(){return[]}},5062:a=>{a.exports=function(){return!1}},75:function(a){(function(){var u,s,c,f,d,p;typeof performance<"u"&&performance!==null&&performance.now?a.exports=function(){return performance.now()}:typeof process<"u"&&process!==null&&process.hrtime?(a.exports=function(){return(u()-d)/1e6},s=process.hrtime,f=(u=function(){var m;return 1e9*(m=s())[0]+m[1]})(),p=1e9*process.uptime(),d=f-p):Date.now?(a.exports=function(){return Date.now()-c},c=Date.now()):(a.exports=function(){return new Date().getTime()-c},c=new Date().getTime())}).call(this)},4087:(a,u,s)=>{for(var c=s(75),f=typeof window>"u"?s.g:window,d=["moz","webkit"],p="AnimationFrame",m=f["request"+p],w=f["cancel"+p]||f["cancelRequest"+p],y=0;!m&&y<d.length;y++)m=f[d[y]+"Request"+p],w=f[d[y]+"Cancel"+p]||f[d[y]+"CancelRequest"+p];if(!m||!w){var v=0,h=0,g=[];m=function(E){if(g.length===0){var _=c(),S=Math.max(0,16.666666666666668-(_-v));v=S+_,setTimeout(function(){var x=g.slice(0);g.length=0;for(var R=0;R<x.length;R++)if(!x[R].cancelled)try{x[R].callback(v)}catch(z){setTimeout(function(){throw z},0)}},Math.round(S))}return g.push({handle:++h,callback:E,cancelled:!1}),h},w=function(E){for(var _=0;_<g.length;_++)g[_].handle===E&&(g[_].cancelled=!0)}}a.exports=function(E){return m.call(f,E)},a.exports.cancel=function(){w.apply(f,arguments)},a.exports.polyfill=function(E){E||(E=f),E.requestAnimationFrame=m,E.cancelAnimationFrame=w}},8156:a=>{a.exports=n}},o={};function i(a){var u=o[a];if(u!==void 0)return u.exports;var s=o[a]={id:a,loaded:!1,exports:{}};return r[a].call(s.exports,s,s.exports,i),s.loaded=!0,s.exports}i.n=a=>{var u=a&&a.__esModule?()=>a.default:()=>a;return i.d(u,{a:u}),u},i.d=(a,u)=>{for(var s in u)i.o(u,s)&&!i.o(a,s)&&Object.defineProperty(a,s,{enumerable:!0,get:u[s]})},i.g=function(){if(typeof globalThis=="object")return globalThis;try{return this||new Function("return this")()}catch{if(typeof window=="object")return window}}(),i.o=(a,u)=>Object.prototype.hasOwnProperty.call(a,u),i.nmd=a=>(a.paths=[],a.children||(a.children=[]),a);var l={};return(()=>{i.d(l,{default:()=>_});var a=i(8156),u=i.n(a),s=i(7403),c=i(8446),f=i.n(c);function d(S){return d=typeof Symbol=="function"&&typeof Symbol.iterator=="symbol"?function(x){return typeof x}:function(x){return x&&typeof Symbol=="function"&&x.constructor===Symbol&&x!==Symbol.prototype?"symbol":typeof x},d(S)}function p(S,x){if(!(S instanceof x))throw new TypeError("Cannot call a class as a function")}function m(S,x){for(var R=0;R<x.length;R++){var z=x[R];z.enumerable=z.enumerable||!1,z.configurable=!0,"value"in z&&(z.writable=!0),Object.defineProperty(S,z.key,z)}}function w(S,x){return w=Object.setPrototypeOf?Object.setPrototypeOf.bind():function(R,z){return R.__proto__=z,R},w(S,x)}function y(S,x){if(x&&(d(x)==="object"||typeof x=="function"))return x;if(x!==void 0)throw new TypeError("Derived constructors may only return object or undefined");return v(S)}function v(S){if(S===void 0)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return S}function h(S){return h=Object.setPrototypeOf?Object.getPrototypeOf.bind():function(x){return x.__proto__||Object.getPrototypeOf(x)},h(S)}function g(S,x,R){return x in S?Object.defineProperty(S,x,{value:R,enumerable:!0,configurable:!0,writable:!0}):S[x]=R,S}var E=function(S){(function(W,b){if(typeof b!="function"&&b!==null)throw new TypeError("Super expression must either be null or a function");W.prototype=Object.create(b&&b.prototype,{constructor:{value:W,writable:!0,configurable:!0}}),Object.defineProperty(W,"prototype",{writable:!1}),b&&w(W,b)})(J,S);var x,R,z,j,H=(z=J,j=function(){if(typeof Reflect>"u"||!Reflect.construct||Reflect.construct.sham)return!1;if(typeof Proxy=="function")return!0;try{return Boolean.prototype.valueOf.call(Reflect.construct(Boolean,[],function(){})),!0}catch{return!1}}(),function(){var W,b=h(z);if(j){var q=h(this).constructor;W=Reflect.construct(b,arguments,q)}else W=b.apply(this,arguments);return y(this,W)});function J(){var W;p(this,J);for(var b=arguments.length,q=new Array(b),X=0;X<b;X++)q[X]=arguments[X];return g(v(W=H.call.apply(H,[this].concat(q))),"state",{instance:null}),W}return x=J,(R=[{key:"componentDidMount",value:function(){var W=this,b=new s.default(this.typewriter,this.props.options);this.setState({instance:b},function(){var q=W.props.onInit;q&&q(b)})}},{key:"componentDidUpdate",value:function(W){f()(this.props.options,W.options)||this.setState({instance:new s.default(this.typewriter,this.props.options)})}},{key:"componentWillUnmount",value:function(){this.state.instance&&this.state.instance.stop()}},{key:"render",value:function(){var W=this,b=this.props.component;return u().createElement(b,{ref:function(q){return W.typewriter=q},className:"Typewriter","data-testid":"typewriter-wrapper"})}}])&&m(x.prototype,R),Object.defineProperty(x,"prototype",{writable:!1}),J}(a.Component);E.defaultProps={component:"div"};const _=E})(),l.default})())})(x1);const _1=ws(gs),mh=N.forwardRef(({bsPrefix:e,variant:t,animation:n="border",size:r,as:o="div",className:i,...l},a)=>{e=Hn(e,"spinner");const u=`${e}-${n}`;return $(o,{ref:a,...l,className:Sn(i,u,r&&`${u}-${r}`,t&&`text-${t}`)})});mh.displayName="Spinner";const k1=()=>{const{user:e}=N.useContext(gl),[t,n]=N.useState(""),[r,o]=N.useState(""),[i,l]=N.useState(!1),[a,u]=N.useState(!1),s=()=>{l(d=>!d)};return xe(Wh,{children:[xe("div",{className:"quote_input",children:[$("form",{onSubmit:async d=>{d.preventDefault(),u(!0);const p=await Lg(r,e.email);n(p.data),u(!1),s()},children:$("input",{placeholder:"Enter a quote here and get three matching book recommendations.",value:r,onChange:d=>o(d.target.value)})}),a?$("div",{className:"spinner_container",children:$(mh,{animation:"border",role:"status",className:"custom_spinner",children:$("span",{className:"visually-hidden",children:"Loading..."})})}):t?$("div",{className:"rec_display",children:$(_1,{options:{strings:[(d=>{if(d)return d.split(`
 `).join("<br/>")})(t.data)],autoStart:!0,loop:!0,delay:10,cursor:"",html:!1,deleteSpeed:1e9}})}):null]}),$(S1,{user:e,update:i,triggerUpdate:s})]})},C1=()=>{const[e,t]=N.useState(""),[n,r]=N.useState(""),{setUser:o}=N.useContext(gl),i=Bp();return xe("form",{onSubmit:async a=>{a.preventDefault(),await Og(e,n,o)==!1?(t(""),r(""),alert("Invalid credentials")):i("/main/")},children:[$("input",{placeholder:"email",value:e,onChange:a=>t(a.target.value)}),$("input",{placeholder:"password",type:"password",value:n,onChange:a=>r(a.target.value)}),$("button",{class:"btn btn-outline-secondary",type:"submit",children:xe("svg",{xmlns:"http://www.w3.org/2000/svg",width:"32",height:"32",fill:"currentColor",className:"bi bi-box-arrow-in-right",viewBox:"0 0 16 16",children:[$("path",{fillRule:"evenodd",d:"M6 3.5a.5.5 0 0 1 .5-.5h8a.5.5 0 0 1 .5.5v9a.5.5 0 0 1-.5.5h-8a.5.5 0 0 1-.5-.5v-2a.5.5 0 0 0-1 0v2A1.5 1.5 0 0 0 6.5 14h8a1.5 1.5 0 0 0 1.5-1.5v-9A1.5 1.5 0 0 0 14.5 2h-8A1.5 1.5 0 0 0 5 3.5v2a.5.5 0 0 0 1 0v-2z"}),$("path",{fillRule:"evenodd",d:"M11.854 8.354a.5.5 0 0 0 0-.708l-3-3a.5.5 0 1 0-.708.708L10.293 7.5H1.5a.5.5 0 0 0 0 1h8.793l-2.147 2.146a.5.5 0 0 0 .708.708l3-3z"})]})})]})},R1=()=>{const[e,t]=N.useState(""),[n,r]=N.useState("");return xe("form",{onSubmit:async i=>{i.preventDefault();const l=await Pg(e,n);console.log(l),l===!1?(t(""),r(""),alert("Invalid email.")):(t(""),r(""),alert("Valid signup. Proceed to sign in."))},style:{display:"flex",flexDirection:"column"},children:[$("input",{placeholder:"Email",value:e,onChange:i=>t(i.target.value)}),$("input",{placeholder:"Password",type:"password",value:n,onChange:i=>r(i.target.value)}),$("button",{style:{alignSelf:"center",width:"60%"},class:"btn btn-secondary",type:"submit",children:"Sign Up"})]})},P1=()=>{const[e,t]=N.useState(!1);return $("div",{className:"auth_container",children:xe("div",{className:"auth_box",children:[e?$(C1,{}):$(R1,{}),xe("span",{style:{marginTop:"10px"},children:[e?"Don't have an account?":"Already have an account?"," ",$("button",{class:"btn btn-secondary",onClick:()=>{t(r=>!r)},children:e?"Sign Up":"Sign In"})]})]})})},O1=f0([{path:"/",element:$(Ug,{}),children:[{index:!0,element:$(P1,{})},{path:"main/",element:$(k1,{})}]}]);fa.createRoot(document.getElementById("root")).render($(yt.StrictMode,{children:$(i0,{router:O1})}));
-//# sourceMappingURL=index-cdf5ff2a.js.map
+//# sourceMappingURL=index-36d96ced.js.map
