+++ static/index.html
@@ -8,7 +8,7 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Cinzel+Decorative&family=Cormorant+Infant&family=Martel&display=swap" rel="stylesheet">
     <title>ChoiceReads</title>
-    <script type="module" crossorigin src="/static/assets/index-cdf5ff2a.js"></script>
+    <script type="module" crossorigin src="/static/assets/index-36d96ced.js"></script>
     <link rel="stylesheet" href="/static/assets/index-6b110d0c.css">
   </head>
   <body>
